### 2024-03-08 #6290 @boisgera
> @boisgera thank you for agreeing to be editor. I have realized that I want to make a slight change to some of my code to improve usability. Am I allowed to do so or will it interfere with the review process?

Hi @Sm00thix,

Sure, go ahead.
