+++ ikpls/jax_ikpls_alg_2.py
@@ -14,17 +14,42 @@ class PLS(PLSBase):
 
     Parameters
     ----------
-    reverse_differentiable: bool, optional (default=False). Whether to make the implementation end-to-end differentiable. The differentiable version is slightly slower. Results among the two versions are identical.
+    center : bool, optional, default=True
+        Whether to center the data before fitting. If True, then the mean of the training data is subtracted from the data. If False, then the data is assumed to be already centered.
 
-    verbose : bool, optional (default=False). If True, each sub-function will print when it will be JIT compiled. This can be useful to track if recompilation is triggered due to passing inputs with different shapes.
+    scale : bool, optional, default=True
+        Whether to scale the data before fitting. If True, then the data is scaled using Bessel's correction for the unbiased estimate of the sample standard deviation. If False, then the data is assumed to be already scaled.
+
+    copy : bool, optional, default=True
+        Whether to copy `X` and `Y` in fit before potentially applying centering and scaling. If True, then the data is copied before fitting. If False, and `dtype` matches the type of `X` and `Y`, then centering and scaling is done inplace, modifying both arrays.
+
+    dtype : jnp.float_, optional, default=jnp.float64
+        The float datatype to use in computation of the PLS algorithm. Using a lower precision than float64 will yield significantly worse results when using an increasing number of components due to propagation of numerical errors.
+
+    reverse_differentiable: bool, optional, default=False
+        Whether to make the implementation end-to-end differentiable. The differentiable version is slightly slower. Results among the two versions are identical.
+
+    verbose : bool, optional, default=False
+        If True, each sub-function will print when it will be JIT compiled. This can be useful to track if recompilation is triggered due to passing inputs with different shapes.
     """
 
     def __init__(
-        self, reverse_differentiable: bool = False, verbose: bool = False
+        self,
+        center: bool = True,
+        scale: bool = True,
+        copy: bool = True,
+        dtype: jnp.float_ = jnp.float64,
+        reverse_differentiable: bool = False,
+        verbose: bool = False,
     ) -> None:
-        name = "Improved Kernel PLS Algorithm #2"
+        self.name = "Improved Kernel PLS Algorithm #2"
         super().__init__(
-            name=name, reverse_differentiable=reverse_differentiable, verbose=verbose
+            center=center,
+            scale=scale,
+            copy=copy,
+            dtype=dtype,
+            reverse_differentiable=reverse_differentiable,
+            verbose=verbose,
         )
 
     def _get_initial_matrices(
@@ -220,10 +245,10 @@ class PLS(PLSBase):
         Parameters
         ----------
         X : Array of shape (N, K)
-            Predictor variables. Its dtype will be converted to float64 for reliable results.
+            Predictor variables.
 
-        Y : Array of shape (N, M)
-            Response variables. Its dtype will be converted to float64 for reliable results.
+        Y : Array of shape (N, M) or (N,)
+            Response variables.
 
         A : int
             Number of components in the PLS model.
@@ -245,6 +270,18 @@ class PLS(PLSBase):
         R : Array of shape (K, A)
             PLS weights matrix to compute scores T directly from original X.
 
+        X_mean : Array of shape (1, K) or None
+            Mean of X. If centering is not performed, this is None.
+
+        Y_mean : Array of shape (1, M) or None
[…]
