+++ time_pls.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "-model",
         type=str,
-        help="Model to use. Must be either 'sk', 'np1', 'np2', 'jax1', 'jax2', 'diffjax1', 'diffjax2'.",
+        help="Model to use. Must be either 'sk', 'np1', 'np2', 'fastnp1, 'fastnp2', 'jax1', 'jax2', 'diffjax1', 'diffjax2'.",
     )
     parser.add_argument("-n_components", type=int, help="Number of components to use.")
     parser.add_argument(
@@ -45,18 +45,18 @@ if __name__ == "__main__":
     X, Y = gen_random_data(n, k, m)
     if "jax" in model:
         if model == "jax1":
-            pls = JAX_PLS_Alg_1()
+            pls = JAX_PLS_Alg_1(center=False, scale=False)
             name = "JAX Improved Kernel PLS Algorithm #1"
         elif model == "jax2":
-            pls = JAX_PLS_Alg_2()
+            pls = JAX_PLS_Alg_2(center=False, scale=False)
             name = "JAX Improved Kernel PLS Algorithm #2"
         elif model == "diffjax1":
-            pls = JAX_PLS_Alg_1(reverse_differentiable=True)
+            pls = JAX_PLS_Alg_1(center=False, scale=False, reverse_differentiable=True)
             name = (
                 "JAX Improved Kernel PLS Algorithm #1 (backwards mode differentiable)"
             )
         elif model == "diffjax2":
-            pls = JAX_PLS_Alg_2(reverse_differentiable=True)
+            pls = JAX_PLS_Alg_2(center=False, scale=False, reverse_differentiable=True)
             name = (
                 "JAX Improved Kernel PLS Algorithm #2 (backwards mode differentiable)"
             )
@@ -80,18 +80,18 @@ if __name__ == "__main__":
             fit_params = {} if n_splits == 1 else None
             name = "scikit-learn NIPALS"
         elif model == "np1":
-            pls = NP_PLS(algorithm=1)
+            pls = NP_PLS(center=False, scale=False, algorithm=1)
             fit_params = {"A": n_components}
             name = "NumPy Improved Kernel PLS Algorithm #1"
         elif model == "np2":
-            pls = NP_PLS(algorithm=2)
+            pls = NP_PLS(center=False, scale=False, algorithm=2)
             fit_params = {"A": n_components}
             name = "NumPy Improved Kernel PLS Algorithm #2"
         elif model == "fastnp1":
-            pls = NP_PLS_FCV(algorithm=1)
+            pls = NP_PLS_FCV(center=False, scale=False, algorithm=1)
             name = "NumPy Improved Kernel PLS Algorithm #1 (fast cross-validation)"
         elif model == "fastnp2":
-            pls = NP_PLS_FCV(algorithm=2)
+            pls = NP_PLS_FCV(center=False, scale=False, algorithm=2)
             name = "NumPy Improved Kernel PLS Algorithm #2 (fast cross-validation)"
         else:
             raise ValueError(
