+++ ikpls/jax_ikpls_base.py
@@ -19,20 +19,38 @@ class PLSBase(abc.ABC):
 
     Parameters
     ----------
-    reverse_differentiable: bool, optional (default=False). Whether to make the implementation end-to-end differentiable. The differentiable version is slightly slower. Results among the two versions are identical.
+    center : bool, optional, default=True
+        Whether to center the data before fitting. If True, then the mean of the training data is subtracted from the data. If False, then the data is assumed to be already centered.
 
-    name : str, optional (default=\"Improved Kernel PLS Algorithm\"). Assigns a name to the instance of the class.
+    scale : bool, optional, default=True
+        Whether to scale the data before fitting. If True, then the data is scaled using Bessel's correction for the unbiased estimate of the sample standard deviation. If False, then the data is assumed to be already scaled.
 
-    verbose : bool, optional (default=False). If True, each sub-function will print when it will be JIT compiled. This can be useful to track if recompilation is triggered due to passing inputs with different shapes.
+    copy : bool, optional, default=True
+        Whether to copy `X` and `Y` in fit before potentially applying centering and scaling. If True, then the data is copied before fitting. If False, and `dtype` matches the type of `X` and `Y`, then centering and scaling is done inplace, modifying both arrays.
+
+    dtype : jnp.float, optional, default=jnp.float64
+        The float datatype to use in computation of the PLS algorithm. Using a lower precision than float64 will yield significantly worse results when using an increasing number of components due to propagation of numerical errors.
+
+    reverse_differentiable: bool, optional, default=False
+        Whether to make the implementation end-to-end differentiable. The differentiable version is slightly slower. Results among the two versions are identical.
+
+    verbose : bool, optional, default=False
+        If True, each sub-function will print when it will be JIT compiled. This can be useful to track if recompilation is triggered due to passing inputs with different shapes.
     """
 
     def __init__(
         self,
-        name: str = "Improved Kernel PLS Algorithm",
+        center: bool = True,
+        scale: bool = True,
+        copy: bool = True,
+        dtype: jnp.float_ = jnp.float64,
         reverse_differentiable: bool = False,
         verbose: bool = False,
     ) -> None:
-        self.name = name
+        self.center = center
+        self.scale = scale
+        self.copy = copy
+        self.dtype = dtype
         self.reverse_differentiable = reverse_differentiable
         self.verbose = verbose
 
@@ -93,6 +111,136 @@ class PLSBase(abc.ABC):
         b = b_last + r @ q.T
         return b
 
+    @partial(jax.jit, static_argnums=0)
+    def _initialize_input_matrices(self, X, Y):
+        """
+        Initialize the input matrices used in the PLS algorithm.
+
+        Parameters
+        ----------
+        X : Array of shape (N, K)
+            Predictor variables matrix.
+
+        Y : Array of shape (N, M) or (N,)
+            Response variables matrix.
+        """
+        X = jnp.asarray(X, dtype=self.dtype)
+        Y = jnp.asarray(Y, dtype=self.dtype)
+
+        if Y.ndim == 1:
+            Y = Y.reshape(-1, 1)
+        return X, Y
+
+    @partial(jax.jit, static_argnums=0)
+    def get_means(self, X, Y):
+        """
+        Get the mean of the input matrices.
+
+        Parameters
+        ----------
+        X : Array of shape (N, K)
+            Predictor variables matrix.
+
[…]

[…]
