### 2024-03-10 #6290 @Sm00thix
Thank you, @boisgera. I have now updated the main branch with my desired changes. Specifically, I included the option to center and/or scale the input data directly in the models. I have updated the documentation, tests, and examples to reflect the changes. I have also released the new version of the package on PyPI. The paper remains unchanged. Now, I am ready to continue with the peer review process.
