### 2024-03-10 ca56a34
Updated docs and added JOSS badge
