### 2024-03-10 e7a1fc3
Updated docs, type annotations, and tests.
