+++ ikpls/jax_ikpls_alg_1.py
@@ -1,9 +1,10 @@
 from functools import partial
-from typing import Tuple, Union
+from typing import Tuple
 
 import jax
 import jax.numpy as jnp
 from jax.experimental import host_callback
+from jax.typing import ArrayLike, DTypeLike
 
 from ikpls.jax_ikpls_base import PLSBase
 
@@ -23,7 +24,7 @@ class PLS(PLSBase):
     copy : bool, optional, default=True
         Whether to copy `X` and `Y` in fit before potentially applying centering and scaling. If True, then the data is copied before fitting. If False, and `dtype` matches the type of `X` and `Y`, then centering and scaling is done inplace, modifying both arrays.
 
-    dtype : jnp.float_, optional, default=jnp.float64
+    dtype : DTypeLike, optional, default=jnp.float64
         The float datatype to use in computation of the PLS algorithm. Using a lower precision than float64 will yield significantly worse results when using an increasing number of components due to propagation of numerical errors.
 
     reverse_differentiable: bool, optional, default=False
@@ -38,7 +39,7 @@ class PLS(PLSBase):
         center: bool = True,
         scale: bool = True,
         copy: bool = True,
-        dtype: jnp.float_ = jnp.float64,
+        dtype: DTypeLike = jnp.float64,
         reverse_differentiable: bool = False,
         verbose: bool = False,
     ) -> None:
@@ -55,9 +56,7 @@ class PLS(PLSBase):
     @partial(jax.jit, static_argnums=(0, 1, 2, 3, 4))
     def _get_initial_matrices(
         self, A: int, K: int, M: int, N: int
-    ) -> Tuple[
-        jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray
-    ]:
+    ) -> Tuple[jax.Array, jax.Array, jax.Array, jax.Array, jax.Array, jax.Array]:
         """
         Initialize the matrices and arrays needed for the PLS algorithm. This method is part of the PLS fitting process.
 
@@ -102,7 +101,7 @@ class PLS(PLSBase):
         return B, W, P, Q, R, T
 
     @partial(jax.jit, static_argnums=(0,))
-    def _step_1(self, X: jnp.ndarray, Y: jnp.ndarray) -> jnp.ndarray:
+    def _step_1(self, X: jax.Array, Y: jax.Array) -> jax.Array:
         """
         Perform the first step of Improved Kernel PLS Algorithm #1.
 
@@ -124,7 +123,7 @@ class PLS(PLSBase):
         return self._compute_initial_XTY(X.T, Y)
 
     @partial(jax.jit, static_argnums=(0,))
-    def _step_4(self, X: jnp.ndarray, XTY: jnp.ndarray, r: jnp.ndarray):
+    def _step_4(self, X: jax.Array, XTY: jax.Array, r: jax.Array):
         """
         Perform the fourth step of Improved Kernel PLS Algorithm #1.
 
@@ -171,16 +170,14 @@ class PLS(PLSBase):
         self,
         A: int,
         i: int,
-        X: jnp.ndarray,
-        XTY: jnp.ndarray,
+        X: jax.Array,
+        XTY: jax.Array,
         M: int,
         K: int,
-        P: jnp.ndarray,
-        R: jnp.ndarray,
+        P: jax.Array,
+        R: jax.Array,
         reverse_differentiable: bool,
-    ) -> Tuple[
-        jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray, jnp.ndarray
-    ]:
+    ) -> Tuple[jax.Array, jax.Array, jax.Array, jax.Array, jax.Array, jax.Array]:
         """
[…]
