+++ tests/test_ikpls.py
@@ -869,6 +869,75 @@ class TestClass:
             rtol=1e-5,
         )  # PLS1 is very numerically stable for protein.
 
+        Y = (
+            Y.squeeze()
+        )  # Remove the singleton dimension and check that the predictions are consistent.
+        assert Y.ndim == 1
+        (
+            sk_pls,
+            sk_B,
+            np_pls_alg_1,
+            np_pls_alg_2,
+            jax_pls_alg_1,
+            jax_pls_alg_2,
+            diff_jax_pls_alg_1,
+            diff_jax_pls_alg_2,
+        ) = self.fit_models(X=X, Y=Y, n_components=n_components)
+
+        self.check_cpu_gpu_equality(
+            np_pls_alg_1=np_pls_alg_1,
+            np_pls_alg_2=np_pls_alg_2,
+            jax_pls_alg_1=jax_pls_alg_1,
+            jax_pls_alg_2=jax_pls_alg_2,
+            diff_jax_pls_alg_1=diff_jax_pls_alg_1,
+            diff_jax_pls_alg_2=diff_jax_pls_alg_2,
+        )
+
+        self.check_equality_properties(
+            np_pls_alg_1=np_pls_alg_1,
+            jax_pls_alg_1=jax_pls_alg_1,
+            diff_jax_pls_alg_1=diff_jax_pls_alg_1,
+            X=X,
+            atol=1e-1,
+            rtol=1e-5,
+        )
+        self.check_orthogonality_properties(
+            np_pls_alg_1=np_pls_alg_1,
+            np_pls_alg_2=np_pls_alg_2,
+            jax_pls_alg_1=jax_pls_alg_1,
+            jax_pls_alg_2=jax_pls_alg_2,
+            diff_jax_pls_alg_1=diff_jax_pls_alg_1,
+            diff_jax_pls_alg_2=diff_jax_pls_alg_2,
+            atol=1e-1,
+            rtol=0,
+        )
+
+        self.check_regression_matrices(
+            sk_B=sk_B,
+            np_pls_alg_1=np_pls_alg_1,
+            np_pls_alg_2=np_pls_alg_2,
+            jax_pls_alg_1=jax_pls_alg_1,
+            jax_pls_alg_2=jax_pls_alg_2,
+            diff_jax_pls_alg_1=diff_jax_pls_alg_1,
+            diff_jax_pls_alg_2=diff_jax_pls_alg_2,
+            atol=1e-8,
+            rtol=6e-5,
+        )
+
+        self.check_predictions(
+            sk_pls=sk_pls,
+            sk_B=sk_B,
+            np_pls_alg_1=np_pls_alg_1,
+            np_pls_alg_2=np_pls_alg_2,
+            jax_pls_alg_1=jax_pls_alg_1,
+            jax_pls_alg_2=jax_pls_alg_2,
+            diff_jax_pls_alg_1=diff_jax_pls_alg_1,
+            diff_jax_pls_alg_2=diff_jax_pls_alg_2,
+            X=X,
+            atol=1e-8,
+            rtol=1e-5,
+        )  # PLS1 is very numerically stable for protein.
+
     def test_pls_2_m_less_k(self):
         """
         Description
@@ -1688,27 +1757,18 @@ class TestClass:
         AssertionError
             If the warnings for weights being close to zero are not raised for all implementations.
         """
-        ## Taken from self.fit_models to check each individual algorithm for early stopping.
-        x_mean = X.mean(axis=0)
-        X -= x_mean
-        y_mean = Y.mean(axis=0)
-        Y -= y_mean
-        x_std = X.std(axis=0, ddof=1)
-        x_std[x_std == 0.0] = 1.0
[…]
