+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ikpls"
-version = "1.1.2"
+version = "1.1.3"
 description = ""
 authors = ["Sm00thix <oleemail@icloud.com>"]
 maintainers = ["Sm00thix <oleemail@icloud.com>"]
