### 2024-03-10 bd03285
Fixed typeguard installation in test action
