+++ .github/actions/test/action.yml
@@ -19,7 +19,6 @@ runs:
         python3 -m pip install flake8 pytest
         python3 -m pip install build --user
         pip3 install -U "jax[cpu]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
-        pip3 install pytest pytest-cov
-        pip3 install pytest typeguard
+        pip3 install pytest pytest-cov typeguard
         python3 -m pytest tests --doctest-modules --junitxml=junit/test-results.xml --cov=ikpls/ --cov-report=xml --cov-report=html --typeguard-packages=ikpls/
       shell: bash 
