### 2024-03-10 #6290 @Sm00thix
Hi again,
I made some additional updates to the code. I updated the type annotations and added typeguard to the tests to verify the annotations. The documentation and package have been updated accordingly.
