### 2024-03-14 #6290 @boisgera
> Hi again, I made some additional updates to the code. I updated the type annotations and added typeguard to the tests to verify the annotations. The documentation and package have been updated accordingly.

OK, great! I am currently contacting potential reviewers for your paper ; I'll let you know as soon 2-3 of them are willing to make the review.
