### 2024-03-24 #6533 @editorialbot
[REVIEW]: IKPLS: Improved Kernel Partial Least Squares and Fast Cross-Validation Algorithms for Python with CPU and GPU Implementations Using NumPy and JAX
**Submitting author:** @Sm00thix (<a href="http://orcid.org/0000-0002-7906-4589">Ole-Christian Galbo Engstrøm</a>)
**Repository:** https://github.com/Sm00thix/IKPLS
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.4
