### 2024-03-25 Sm00thix/IKPLS#9 @parmentelat
pylint-friendliness & very long lines
Hey there

as part of the review initiated in https://github.com/openjournals/joss-reviews/issues/6533, I am having a first stab at this repo :)

a very first and general comment that I have about the Python code relates to pylint assessment:

- particularly the "line-too-long" messages that hinder smooth reading (I take it your editor wraps long lines ? I have seen some of them reaching 350+ characters, this is not a good practice !)
- and more generally a rather poor score out of pylint, which gives me
  - a score of 0/10 on the `ikps` folder
  - a score of 5.14/10 on the `examples` folder

given the heavy use of 'math-like' variables, I guess it is acceptable to allow for the `invalid-name` issues
(for example if I do allow that on one source file (`ikpls/jax_ikpls_alg_2.py`) its score rises from 0/10 to 4.15/10)

so i guess my proposal would be

- to achieve a better score by pylint in a *best-effort* way
- by at least sticking to a strict 88-columns max line width
- and solving - or at least starting to solve - the other issues, apart from the "invalid-name" ones that are endemic but acceptable due to the variable naming scheme
