### 2024-03-26 243fa8b
Removed JAX_ENABLE_X64 environment variable from worklow. Instead added a jax.config update in the test_ikpls.py file to allow JAX using 64 bit floats.
