+++ .github/workflows/workflow.yml
@@ -22,9 +22,7 @@ jobs:
       - uses: ./.github/actions/test
 
   build_package:
-    runs-on: ubuntu-latest  
-    env:
-      JAX_ENABLE_X64: True
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - uses: ./.github/actions/build
