### 2024-03-26 0c59442
Added commenet describing the meaning of updating the JAX config
