### 2024-03-27 f3f95eb
Added README to ./paper to document how to generate timings.png and how to (re)produce benchmarks.
