+++ paper/time_pls.py
@@ -1,0 +1,126 @@
+import argparse
+
+from ikpls.fast_cross_validation.numpy_ikpls import PLS as NP_PLS_FCV
+from ikpls.jax_ikpls_alg_1 import PLS as JAX_PLS_Alg_1
+from ikpls.jax_ikpls_alg_2 import PLS as JAX_PLS_Alg_2
+from ikpls.numpy_ikpls import PLS as NP_PLS
+from timings.timings import (
+    SK_PLS_All_Components,
+    cross_val_cpu_pls,
+    cross_val_gpu_pls,
+    fast_cross_val_cpu_pls,
+    gen_random_data,
+    single_fit_cpu_pls,
+    single_fit_gpu_pls,
+)
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-model",
+        type=str,
+        help="Model to use. Must be either 'sk', 'np1', 'np2', 'fastnp1, 'fastnp2', 'jax1', 'jax2', 'diffjax1', 'diffjax2'. Note that fastnp1 and fastnp2 only support cross-validation and not single fits.",
+    )
+    parser.add_argument("-n_components", type=int, help="Number of components to use.")
+    parser.add_argument(
+        "-n_splits", type=int, help="Number of splits to use in cross-validation. If 1, will perform a single fit."
+    )
+    parser.add_argument(
+        "-n_jobs",
+        type=int,
+        help="Number of parallel jobs to use. A value of -1 will use all available cores. Not used for JAX implementations as it is assumed these will run on a TPU or GPU.",
+    )
+    parser.add_argument("-n", type=int, help="Number of samples to generate.")
+    parser.add_argument("-k", type=int, help="Number of features to generate.")
+    parser.add_argument("-m", type=int, help="Number of targets to generate.")
+    args = parser.parse_args()
+    config = vars(args)
+    model = config["model"]
+    n_components = config["n_components"]
+    n_splits = config["n_splits"]
+    n = config["n"]
+    k = config["k"]
+    m = config["m"]
+
+    X, Y = gen_random_data(n, k, m)
+    if "jax" in model:
+        if model == "jax1":
+            pls = JAX_PLS_Alg_1(center=False, scale=False)
+            name = "JAX Improved Kernel PLS Algorithm #1"
+        elif model == "jax2":
+            pls = JAX_PLS_Alg_2(center=False, scale=False)
+            name = "JAX Improved Kernel PLS Algorithm #2"
+        elif model == "diffjax1":
+            pls = JAX_PLS_Alg_1(center=False, scale=False, reverse_differentiable=True)
+            name = (
+                "JAX Improved Kernel PLS Algorithm #1 (backwards mode differentiable)"
+            )
+        elif model == "diffjax2":
+            pls = JAX_PLS_Alg_2(center=False, scale=False, reverse_differentiable=True)
+            name = (
+                "JAX Improved Kernel PLS Algorithm #2 (backwards mode differentiable)"
+            )
+        if n_splits == 1:
+            print(
+                f"Fitting {name} with {n_components} components on {n} samples with {k} features and {m} targets."
+            )
+            time = single_fit_gpu_pls(pls, X, Y, n_components)
+        else:
+            print(
+                f"Fitting {name} with {n_components} components using {n_splits}-fold cross-validation on {n} samples with {k} features and {m} targets."
+            )
+            time = cross_val_gpu_pls(
+                pls, X, Y, n_components, n_splits, show_progress=True
+            )
+        print(f"Time: {time}")
+    else:
+        n_jobs = config["n_jobs"]
+        if model == "sk":
+            pls = SK_PLS_All_Components(n_components=n_components)
[…]
