+++ paper/timings/timings.py
@@ -1,0 +1,475 @@
+from timeit import Timer, default_timer
+
+import jax.numpy as jnp
+import numpy as np
+import numpy.typing as npt
+from numpy import ndarray
+from sklearn.cross_decomposition import PLSRegression as SK_PLS
+from sklearn.model_selection import KFold, cross_validate
+
+
+class SK_PLS_All_Components(SK_PLS):
+    """
+    Description
+    -----------
+    Subclass of sklearn's PLSRegression that stores regression matrices for all possible components during fitting.
+
+    Parameters
+    ----------
+    n_components : int
+        Number of components to use in the PLS fit.
+
+    **kwargs:
+        Additional keyword arguments to pass to the superclass constructor.
+    """
+
+    def __init__(self, n_components, **kwargs):
+        super().__init__(n_components=n_components, **kwargs)
+
+    def fit(
+        self,
+        X,
+        Y,
+    ):  # Override fit method to store regression matrices for all possible number of components. This is MUCH faster than fitting an entirely new model for every single number of components
+        """
+        Description
+        -----------
+        Fit the PLS model and store regression matrices for all possible components.
+
+        Parameters
+        ----------
+        X : Array of shape (N, K)
+            Predictor variables.
+
+        Y : Array of shape (N, M)
+            Response variables.
+        """
+        super().fit(X, Y)
+        B = np.empty(
+            (self.n_components, self.n_features_in_, self.y_loadings_.shape[0])
+        )
+        for i in range(B.shape[0]):
+            B_at_component_i = np.dot(
+                self.x_rotations_[..., : i + 1],
+                self.y_loadings_[..., : i + 1].T,
+            )
+            B[i] = B_at_component_i
+        self.B = B
+
+    def predict(self, X: npt.ArrayLike) -> ndarray:
+        """
+        Description
+        -----------
+        Predict the output for each number of components up to A.
+        This is MUCH faster than calling the predict function
+        n_components number of times with different number of components.
+
+        Parameters
+        ----------
+        X : Array of shape (N, K)
+            Predictor variables.
+
+        Returns:
+        Y_pred : Array of shape (A, N, M)
+            Predicted response variables for each number of components up to A.
+        """
+        return (X - self._x_mean) / self._x_std @ self.B + self.intercept_
+
+
+def gen_random_data(N, K, M):
+    """
+    Description
+    -----------
+    Generate random data for testing.
+
+    Parameters
+    ----------
+    N : int
+        Number of samples.
+
+    K : int
+        Number of features in X.
+    M : int
+        Number of features in Y.
+
+    Returns
+    -------
+    (X, Y) : Tuple of arrays
+        Randomly generated input X and target Y.
+    """
+    rng = np.random.default_rng(seed=42)
+    X = rng.random((N, K), dtype=np.float64)
+    Y = rng.random((N, M), dtype=np.float64)
+    return X, Y
+
+
+def mse_for_each_target(estimator, X, Y_true, **kwargs):
+    """
+    Description
+    -----------
+    Calculate lowest mean squared error (mse) for each target and the corresponding number of components that minimizes the mse.
+
+    Parameters
+    ----------
+    estimator : Estimator object
+        PLS estimator with a 'predict' method.
+
+    X : Array of shape (N, K)
+        Predictor variables.
+
+    Y_true : Array of shape (N, M)
+        True response variables.
+
+    **kwargs:
[…]
