+++ plot_timings.py
@@ -1,386 +1,0 @@
-import matplotlib.pyplot as plt
-import matplotlib.ticker as mticker
-import numpy as np
-import pandas as pd
-from matplotlib import patches
-from matplotlib.transforms import blended_transform_factory
-
-
-def remove_rows_where_all_values_except_time_are_same(df):
-    df = df.drop_duplicates(subset=df.columns[:-1])
-    return df
-
-
-def get_name_x_t_dict(df, x_name, constants_dict, single_fit_or_loocv):
-    model_names = ["sk", "np1", "np2", "fastnp1", "fastnp2", "jax1", "jax2"]
-    for name, value in constants_dict.items():
-        df = df[df[name] == value]
-    name_x_t_dict = {}
-    for model_name in model_names:
-        sub_df = df[df["model"] == model_name]
-        if single_fit_or_loocv == "single_fit":
-            sub_df = sub_df[sub_df["n_splits"] == 1]
-        elif single_fit_or_loocv == "loocv":
-            sub_df = sub_df[sub_df["n_splits"] != 1]
-        else:
-            raise ValueError(
-                f"single_fit_or_loocv must be 'single_fit' or 'loocv'. but got: {single_fit_or_loocv}"
-            )
-        x = sub_df[x_name].values
-        t = sub_df["time"].values
-        inferred = sub_df["inferred"].values
-        if model_name == "sk":
-            model_name = "scikit-learn NIPALS (CPU)"
-        elif model_name == "np1":
-            model_name = "NumPy IKPLS #1 (CPU)"
-        elif model_name == "np2":
-            model_name = "NumPy IKPLS #2 (CPU)"
-        elif model_name == "jax1":
-            model_name = "JAX IKPLS #1 (GPU)"
-        elif model_name == "jax2":
-            model_name = "JAX IKPLS #2 (GPU)"
-        elif model_name == "fastnp1":
-            model_name = "NumPy IKPLS #1 (fast cross-validation)"
-        elif model_name == "fastnp2":
-            model_name = "NumPy IKPLS #2 (fast cross-validation)"
-        name_x_t_dict[model_name] = {"x": x, "t": t, "inferred": inferred}
-    return name_x_t_dict
-
-
-def plot_timings(
-    ax,
-    name_x_t_dict,
-    xlabel,
-    constants_dict,
-):
-    fixed_points = [1, 60, 3600, 86400, 604800, 2592000, 31536000, 315360000]
-    fixed_points_labels = [
-        "1 second",
-        "1 minute",
-        "1 hour",
-        "1 day",
-        "1 week",
-        "30 days",
-        "365 days",
-        "3650 days",
-    ]
-    min_t = np.inf
-    max_t = -np.inf
-    np1_color = None
-    np2_color = None
-    for name, x_t_dict in name_x_t_dict.items():
-        x = x_t_dict["x"]
-        t = x_t_dict["t"]
-        inferred = x_t_dict["inferred"]
-        sorted_indices = np.argsort(x)
-        x = x[sorted_indices]
-        t = t[sorted_indices]
-        inferred = inferred[sorted_indices]
-        if "fast" in name:
-            linestyle = "--"
-        else:
-            linestyle = "-"
-        if np1_color is not None and "NumPy IKPLS #1" in name:
-            color = np1_color
-            curve = ax.loglog(x, t, linestyle, color=color, label=name)
-        elif np2_color is not None and "NumPy IKPLS #2" in name:
[…]

[…]
