### 2024-03-27 b3783f7
Updated README description to clarify that row-wise preprocessing is trivially correct for the fast cross-validation algorithms
