### 2024-03-29 a4fb99b
Updated examples and README with increased pylint scores
