### 2024-04-01 936c5d6
Updated CONTRIBUTING.rst in response to #10
