### 2024-04-01 dbe1325
Updated CONTRIBUTING.rst in response to #10
