### 2024-04-01 fe392c0
Updated CONTRIBUTING.rst in response to #10
