### 2024-04-01 29fb9c3
Updated CONTRIBUTING.rst in response to #10
