### 2024-04-01 085df9e
Updated CONTRIBUTING.rst in response to #10
