### 2024-04-01 918713d
Updated CONTRIBUTING.rst in response to #10
