+++ CONTRIBUTING.rst
@@ -40,7 +40,7 @@ You want to make some kind of change to the code base
 #. make sure the existing tests still work by running ``python3 -m pytest tests --typeguard-packages=ikpls/``;
 #. add your own tests (if necessary);
 #. update or expand the documentation;
-#. check that you can build the package locally and that it passes twine check. See `Build from source build_from_source`_ for more information;
+#. check that you can build the package locally and that it passes twine check. See `Build from source <#build_from_source>`_ for more information;
 #. `push <http://rogerdudler.github.io/git-guide/>`_ your feature branch to (your fork of) the IKPLS repository on GitHub;
 #. create the pull request, e.g. following the instructions `here <https://help.github.com/articles/creating-a-pull-request/>`__.
 
