### 2024-04-01 b4e7811
Updated CONTRIBUTING.rst in response to #10
