### 2024-04-01 0116482
Updated CONTRIBUTING.rst in response to #10
