### 2024-04-01 d60b44b
Updated CONTRIBUTING.rst in response to #10
