### 2024-04-01 238e9ea
Updated CONTRIBUTING.rst and README.rst in response to #10
