### 2024-04-01 d05b7dd
Updated README.rst in response to #10
