+++ docs/contributing.rst
@@ -1,0 +1,6 @@
+=======================
+Contribution Guidelines
+=======================
+.. toctree::
+
+    CONTRIBUTING.rst
