+++ docs/contributing.rst
@@ -1,6 +1,0 @@
-=======================
-Contribution Guidelines
-=======================
-.. toctree::
-
-    CONTRIBUTING.rst
