### 2024-04-02 80fe4c4
Enabled tests for pull requests again.
