+++ .github/workflows/pull_request.yml
@@ -8,18 +8,18 @@ on:
     branches: [ "main" ]
 
 jobs:
-  # test_package:
-  #   runs-on: ${{ matrix.os }}  
-  #   strategy:  
-  #     fail-fast: false  
-  #     matrix:  
-  #       os: [ubuntu-latest, windows-latest, macos-latest] 
-  #       python-version: ["3.9", "3.10", "3.11", "3.12"]  
-  #   env:  
-  #     JAX_ENABLE_X64: True  
-  #   steps:
-  #     - uses: actions/checkout@v3
-  #     - uses: ./.github/actions/test
+  test_package:
+    runs-on: ${{ matrix.os }}  
+    strategy:  
+      fail-fast: false  
+      matrix:  
+        os: [ubuntu-latest, windows-latest, macos-latest] 
+        python-version: ["3.9", "3.10", "3.11", "3.12"]  
+    env:  
+      JAX_ENABLE_X64: True  
+    steps:
+      - uses: actions/checkout@v3
+      - uses: ./.github/actions/test
 
   build_package:
     runs-on: ubuntu-latest  
