### 2024-04-02 e5ca1ba
Updated test action to explicitly download jax for cpu
