+++ .github/actions/test/action.yml
@@ -15,7 +15,6 @@ runs:
         python3 -m pip install --upgrade pandas
         python3 -m pip install poetry
         poetry install
-        pip3 install -U "jax[cpu]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
         python3 -m pip install flake8 pytest
         pip3 install pytest pytest-cov typeguard
         poetry run pytest tests --doctest-modules --junitxml=junit/test-results.xml --cov=ikpls/ --cov-report=xml --cov-report=html --typeguard-packages=ikpls/
