+++ .github/actions/test/action.yml
@@ -9,13 +9,18 @@ inputs:
 runs:
   using: "composite"
   steps:
-    - name: Test
+    - name: Install Poetry
       run: |
-        python3 -m pip install --upgrade pip
-        python3 -m pip install --upgrade pandas
-        python3 -m pip install poetry
-        poetry install
-        python3 -m pip install flake8 pytest
-        python3 -m pip install pytest pytest-cov typeguard
+        python -m pip install --upgrade pip
+        python -m pip install poetry
+
+    - name: Install project dependencies
+      run: poetry install
+
+    - name: Install additional dependencies for testing
+      run: |
+        poetry add --dev pandas pytest flake8 pytest-cov typeguard
+
+    - name: Run tests
+      run: |
         poetry run pytest tests --doctest-modules --junitxml=junit/test-results.xml --cov=ikpls/ --cov-report=xml --cov-report=html --typeguard-packages=ikpls/
-      shell: bash
