### 2024-04-02 2fddb82
Added bash shell to test action
