+++ .github/actions/test/action.yml
@@ -13,14 +13,18 @@ runs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install poetry
+      shell: bash
 
     - name: Install project dependencies
       run: poetry install
+      shell: bash
 
     - name: Install additional dependencies for testing
       run: |
         poetry add --dev pandas pytest flake8 pytest-cov typeguard
+      shell: bash
 
     - name: Run tests
       run: |
         poetry run pytest tests --doctest-modules --junitxml=junit/test-results.xml --cov=ikpls/ --cov-report=xml --cov-report=html --typeguard-packages=ikpls/
+      shell: bash
