### 2024-04-02 d9c4fee
Force poetry to use Python version from the GitHub action
