+++ .github/workflows/pull_request.yml
@@ -14,11 +14,9 @@ jobs:
       fail-fast: false  
       matrix:  
         os: [ubuntu-latest, windows-latest, macos-latest] 
-        python-version: ["3.9", "3.10", "3.11", "3.12"]  
-    env:  
-      JAX_ENABLE_X64: True  
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: ./.github/actions/test
 
   build_package:
@@ -26,5 +24,5 @@ jobs:
     env:
       JAX_ENABLE_X64: True
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: ./.github/actions/build
