+++ .github/workflows/workflow.yml
@@ -14,24 +14,22 @@ jobs:
       fail-fast: false  
       matrix:  
         os: [ubuntu-latest, windows-latest, macos-latest] 
-        python-version: ["3.9", "3.10", "3.11", "3.12"]  
-    env:  
-      JAX_ENABLE_X64: True  
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: ./.github/actions/test
 
   build_package:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: ./.github/actions/build
 
   publish_package:
     needs: [build_package, test_package]
     runs-on: ubuntu-latest
     steps:  
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: ./.github/actions/publish
     permissions:
       id-token: write
