### 2024-04-02 9e9fdc1
Force poetry to use Python version from the GitHub action
