+++ .github/actions/test/action.yml
@@ -2,7 +2,7 @@ name: Test
 description: Test
 
 inputs:
-  PYTHON_VERSION:
+  python-version:
     description: test
     required: true
 
@@ -17,7 +17,7 @@ runs:
 
     - name: Install project dependencies
       run: |
-        poetry env use ${{ inputs.PYTHON_VERSION }} --path
+        poetry env use ${{ inputs.python-version }}
         poetry install
       shell: bash
 
