### 2024-04-02 857eff4
Setup Python version in test action
