### 2024-04-02 de9494f
Force poetry to use Python version from the GitHub action
