+++ .github/actions/test/action.yml
@@ -21,7 +21,9 @@ runs:
       shell: bash
 
     - name: Install project dependencies
-      run: poetry install
+      run: |
+        poetry env use ${{ inputs.python-version }}
+        poetry install
       shell: bash
 
     - name: Install additional dependencies for testing
