### 2024-04-02 d458afd
Force poetry to use Python version from the GitHub action
