+++ ikpls/jax_ikpls_base.py
@@ -1,3 +1,17 @@
+"""
+Implements an abstract class for partial least-squares regression using Improved Kernel
+PLS by Dayal and MacGregor.
+
+Implementations of concrete classes exist for both Improved Kernel PLS Algorithm #1
+and Improved Kernel PLS Algorithm #2.
+
+For more details, refer to the paper: 
+"Improved Kernel Partial Least Squares Regression" by Dayal and MacGregor.
+
+Author: Ole-Christian Galbo Engstrøm
+E-mail: ole.e@di.ku.dk
+"""
+
 import abc
 import warnings
 from collections.abc import Callable
@@ -14,29 +28,45 @@ from tqdm import tqdm
 
 class PLSBase(abc.ABC):
     """
-    Implements an abstract class for partial least-squares regression using Improved Kernel PLS by Dayal and MacGregor: https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23.
+    Implements an abstract class for partial least-squares regression using Improved
+    Kernel PLS by Dayal and MacGregor:
+    https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23
 
-    Implementations of concrete classes exist for both Improved Kernel PLS Algorithm #1 and Improved Kernel PLS Algorithm #2.
+    Implementations of concrete classes exist for both Improved Kernel PLS Algorithm #1
+    and Improved Kernel PLS Algorithm #2.
 
     Parameters
     ----------
     center : bool, optional, default=True
-        Whether to center the data before fitting. If True, then the mean of the training data is subtracted from the data. If False, then the data is assumed to be already centered.
+        Whether to center the data before fitting. If True, then the mean of the
+        training data is subtracted from the data. If False, then the data is assumed
+        to be already centered.
 
     scale : bool, optional, default=True
-        Whether to scale the data before fitting. If True, then the data is scaled using Bessel's correction for the unbiased estimate of the sample standard deviation. If False, then the data is assumed to be already scaled.
+        Whether to scale the data before fitting. If True, then the data is scaled
+        using Bessel's correction for the unbiased estimate of the sample standard
+        deviation. If False, then the data is assumed to be already scaled.
 
     copy : bool, optional, default=True
-        Whether to copy `X` and `Y` in fit before potentially applying centering and scaling. If True, then the data is copied before fitting. If False, and `dtype` matches the type of `X` and `Y`, then centering and scaling is done inplace, modifying both arrays.
+        Whether to copy `X` and `Y` in fit before potentially applying centering and
+        scaling. If True, then the data is copied before fitting. If False, and `dtype`
+        matches the type of `X` and `Y`, then centering and scaling is done inplace,
+        modifying both arrays.
 
     dtype : jnp.float, optional, default=jnp.float64
-        The float datatype to use in computation of the PLS algorithm. Using a lower precision than float64 will yield significantly worse results when using an increasing number of components due to propagation of numerical errors.
+        The float datatype to use in computation of the PLS algorithm. Using a lower
+        precision than float64 will yield significantly worse results when using an
+        increasing number of components due to propagation of numerical errors.
 
     reverse_differentiable: bool, optional, default=False
-        Whether to make the implementation end-to-end differentiable. The differentiable version is slightly slower. Results among the two versions are identical.
+        Whether to make the implementation end-to-end differentiable. The
+        differentiable version is slightly slower. Results among the two versions are
+        identical.
 
     verbose : bool, optional, default=False
-        If True, each sub-function will print when it will be JIT compiled. This can be useful to track if recompilation is triggered due to passing inputs with different shapes.
[…]
