+++ ikpls/numpy_ikpls.py
@@ -1,3 +1,15 @@
+"""
+Contains the PLS Class which implements partial least-squares regression using Improved
+Kernel PLS by Dayal and MacGregor:
+https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23
+
+The PLS class subclasses scikit-learn's BaseEstimator to ensure compatibility with e.g.
+scikit-learn's cross_validate. It is written using NumPy.
+
+Author: Ole-Christian Galbo Engstrøm
+E-mail: ole.e@di.ku.dk
+"""
+
 import warnings
 from typing import Union
 
@@ -9,7 +21,9 @@ from sklearn.base import BaseEstimator
 
 class PLS(BaseEstimator):
     """
-    Implements partial least-squares regression using Improved Kernel PLS by Dayal and MacGregor: https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23
+    Implements partial least-squares regression using Improved Kernel PLS by Dayal and
+    MacGregor:
+    https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23
 
     Parameters
     ----------
@@ -17,16 +31,25 @@ class PLS(BaseEstimator):
         Whether to use Improved Kernel PLS Algorithm #1 or #2.
 
     center : bool, default=True
-        Whether to center the data before fitting. If True, then the mean of the training data is subtracted from the data. If False, then the data is assumed to be already centered.
+        Whether to center the data before fitting. If True, then the mean of the
+        training data is subtracted from the data. If False, then the data is assumed
+        to be already centered.
 
     scale : bool, default=True
-        Whether to scale the data before fitting. If True, then the data is scaled using Bessel's correction for the unbiased estimate of the sample standard deviation. If False, then the data is assumed to be already scaled.
+        Whether to scale the data before fitting. If True, then the data is scaled
+        using Bessel's correction for the unbiased estimate of the sample standard
+        deviation. If False, then the data is assumed to be already scaled.
 
     copy : bool, default=True
-        Whether to copy `X` and `Y` in fit before potentially applying centering and scaling. If True, then the data is copied before fitting. If False, and `dtype` matches the type of `X` and `Y`, then centering and scaling is done inplace, modifying both arrays.
+        Whether to copy `X` and `Y` in fit before potentially applying centering and
+        scaling. If True, then the data is copied before fitting. If False, and `dtype`
+        matches the type of `X` and `Y`, then centering and scaling is done inplace,
+        modifying both arrays.
 
     dtype : numpy.float, default=numpy.float64
-        The float datatype to use in computation of the PLS algorithm. Using a lower precision than float64 will yield significantly worse results when using an increasing number of components due to propagation of numerical errors.
+        The float datatype to use in computation of the PLS algorithm. Using a lower
+        precision than float64 will yield significantly worse results when using an
+        increasing number of components due to propagation of numerical errors.
 
     Raises
     ------
@@ -107,7 +130,8 @@ class PLS(BaseEstimator):
         Warns
         -----
         UserWarning.
-            If at any point during iteration over the number of components `A`, the residual goes below machine precision for jnp.float64.
+            If at any point during iteration over the number of components `A`, the
+            residual goes below machine precision for np.float64.
         """
         X = np.asarray(X, dtype=self.dtype)
         Y = np.asarray(Y, dtype=self.dtype)
@@ -166,7 +190,8 @@ class PLS(BaseEstimator):
                 norm = la.norm(XTY, ord=2)
[…]
