### 2024-04-03 30064b9
Updated README.rst, paper/README.md and adressed pylint scores in tests/ and paper/
