+++ tests/test_ikpls.py
@@ -1,4 +1,18 @@
-from typing import Callable, Tuple, Union
+"""
+This file contains tests for the ikpls package. Some of the tests are taken directly
+from the test-suite for the scikit-learn PLSRegression class. Most of the tests are
+written for this work specifically.
+
+The tests are designed to check the consistency of the PLS algorithm implementations
+across different shapes of input data. The tests also check for consistency between
+the IKPLS implementations in the ikpls package and the scikit-learn NIPALS
+implementation.
+
+Author: Ole-Christian Galbo Engstrøm
+E-mail: ole.e@di.ku.dk
+"""
+
+from typing import Callable, Tuple
 
 import jax
 import numpy as np
@@ -7,6 +21,8 @@ import pytest
 from jax import numpy as jnp
 from numpy.testing import assert_allclose
 from sklearn.cross_decomposition import PLSRegression as SkPLS
+from sklearn.datasets import load_linnerud
+from sklearn.model_selection import cross_validate
 
 from ikpls.fast_cross_validation.numpy_ikpls import PLS as FastCVPLS
 from ikpls.jax_ikpls_alg_1 import PLS as JAX_Alg_1
@@ -20,6 +36,20 @@ jax.config.update("jax_enable_x64", True)
 
 
 class TestClass:
+    """
+    Class for testing the IKPLS implementation.
+
+    This class contains methods for testing the IKPLS implementation.
+
+    Attributes
+    ----------
+    csv : DataFrame
+        The CSV data containing target values.
+
+    raw_spectra : NDArray[float]
+        The raw spectral data.
+    """
+
     csv = load_data.load_csv()
     raw_spectra = load_data.load_spectra()
 
@@ -201,7 +231,8 @@ class TestClass:
             Relative tolerance for checking equality.
 
         n_good_components : int, optional
-            Number of components to check, or -1 to use all possible number of components.
+            Number of components to check, or -1 to use all possible number of
+            components.
 
         Returns
         -------
@@ -307,7 +338,8 @@ class TestClass:
             Relative tolerance for checking equality.
 
         n_good_components : int, optional
-            Number of components to check, or -1 to use all possible number of components.
+            Number of components to check, or -1 to use all possible number of
+            components.
 
         Returns
         -------
@@ -441,7 +473,8 @@ class TestClass:
             Relative tolerance for checking equality.
 
         n_good_components : int, optional
-            Number of components to check, or -1 to use all possible number of components.
+            Number of components to check, or -1 to use all possible number of
+            components.
 
         Returns
         -------
@@ -525,7 +558,8 @@ class TestClass:
             Relative tolerance for comparing matrix values.
 
         n_good_components : int, optional
-            Number of components to check, or -1 to use all possible number of components.
+            Number of components to check, or -1 to use all possible number of
+            components.
 
         Returns
         -------
@@ -542,7 +576,8 @@ class TestClass:
         # Assume that models are fitted on centered and scaled data
         X = (X - X.mean(axis=0, keepdims=True)) / X.std(axis=0, keepdims=True, ddof=1)
 
-        # X can be reconstructed by multiplying X scores (T) and the transpose of X loadings (P)
+        # X can be reconstructed by multiplying X scores (T) and the transpose of X
+        # loadings (P)
         assert_allclose(
             np.dot(
                 np_pls_alg_1.T[..., :n_good_components],
@@ -625,7 +660,8 @@ class TestClass:
             JAX-based PLS model using algorithm 2 with reverse differentiation.
 
         n_good_components : int, optional
-            Number of components to check, or -1 to use all possible number of components.
+            Number of components to check, or -1 to use all possible number of
[…]

[…]
