+++ paper/time_pls.py
@@ -1,4 +1,5 @@
 import argparse
+import os
 
 from ikpls.fast_cross_validation.numpy_ikpls import PLS as NP_PLS_FCV
 from ikpls.jax_ikpls_alg_1 import PLS as JAX_PLS_Alg_1
@@ -23,7 +24,9 @@ if __name__ == "__main__":
     )
     parser.add_argument("-n_components", type=int, help="Number of components to use.")
     parser.add_argument(
-        "-n_splits", type=int, help="Number of splits to use in cross-validation. If 1, will perform a single fit."
+        "-n_splits",
+        type=int,
+        help="Number of splits to use in cross-validation. If 1, will perform a single fit.",
     )
     parser.add_argument(
         "-n_jobs",
@@ -117,10 +120,16 @@ if __name__ == "__main__":
                 )
         print(f"Time: {time}")
 
+    num_cores = os.cpu_count() if n_jobs == -1 else n_jobs
+
     try:
-        with open("timings/timings.csv", "x") as f:
-            f.write("model,n_components,n_splits,n,k,m,time\n")
-            f.write(f"{model},{n_components},{n_splits},{n},{k},{m},{time}\n")
+        with open("timings/user_timings.csv", "x") as f:
+            f.write("model,n_components,n_splits,n,k,m,time,inferred,njobs\n")
+            f.write(
+                f"{model},{n_components},{n_splits},{n},{k},{m},{time},False,{num_cores}\n"
+            )
     except FileExistsError:
-        with open("timings/timings.csv", "a") as f:
-            f.write(f"{model},{n_components},{n_splits},{n},{k},{m},{time}\n")
+        with open("timings/user_timings.csv", "a") as f:
+            f.write(
+                f"{model},{n_components},{n_splits},{n},{k},{m},{time},False,{num_cores}\n"
+            )
