+++ paper/README.md
@@ -4,17 +4,12 @@
   - Fit the model on every training partition.
   - Compute the mean squared error (MSE) for every PLS component on every validation partition.
 
-## To reproduce timings/timings.png, the figure showing benchmarks in paper.md:
-  - Execute the following command in your terminal:
-
-    ```bash
-    python3 plot_timings.py
-    ```
-
-    This will read the contents of timings/timings.csv to generate timings/timings.png.
+## Original benchmarks
+The benchmarks shown in paper.md originate from timings/timings.png which was generated
+by reading contents from timings/timings.csv which in turn was generated by executing time_pls.py
+on the machine and with the arguments specified in paper.md.
 
 ## To make your own benchmarks
-  - Delete timings/timings.csv or empty its contents to keep your benchmarks distinct from the ones provided.
   - Execute time_pls.py with arguments specifying:
     - The PLS algorithm to use.
     - The number of components to use.
@@ -28,11 +23,14 @@
     python3 time_pls.py -model fastnp2 -n 1000000 -k 500 -m 10 -n_components 30 -n_splits 1000000 -n_jobs -1
     ```
 
-    This will run the experiment and append the result as a line in timings/timings.csv.
-  - After executing the desired benchmarks, execute plot_timings.py to generate timings/timings.png with your benchmark results.
-  - Once a benchmark has finished, you can safely run another, the result of which will be appended to timings/timings.csv
+    This will run the experiment and append the result as a line in timings/user_timings.csv.
+  - After executing the desired benchmarks, execute plot_timings.py to generate timings/user_timings.png with your benchmark results.
+  - Once a benchmark has finished, you can safely run another, the result of which will be appended to timings/user_timings.csv
   - Execute the following command in your terminal to get an overview and description of the different arguments and their meaning:
-
     ```bash
     python3 time_pls.py -h
     ```
+
+### A note on cross-validation splits
+A user can execute time_pls.py with any number of cross-validation splits, `n_splits`, and a record of the time taken to execute the experiment will be written to timings/user_timings.csv.
+However, for plot_timings.py to work as intended, `n_splits` should be either 1 or the same value as `n` for leave-one-out cross-validation. If this is violated - e.g. if `n` is 100 and `n_splits` is 10 - plot_timings.py will incorrectly interpret this entry as a leave-one-out cross-validation entry and show the result in the leave-one-out cross-validation part of timings/user_timings.png.
