### 2024-04-04 f359f6e
Addressed warnings in tests. Addressed errors in tests on MacOS due to too low rtol in PLS1 gradient test. Updated all PLS algorithms to show a weight warning for each call to fit that raises such a warning instead of only showing the first one. Updated tests to ensure that a weight warning is raised for each call to fit with data that should raise a weight warning. Adressed pylint scores.
