+++ .github/actions/test/action.yml
@@ -26,7 +26,7 @@ runs:
 
     - name: Install additional dependencies for testing
       run: |
-        poetry add --dev pandas pytest flake8 pytest-cov typeguard
+        poetry add --group dev pandas pytest flake8 pytest-cov typeguard
       shell: bash
 
     - name: Run tests

[…]
