### 2024-04-05 a5d4995
Adressed warnings in test module from JAX due to os.fork() being called by other PLS implementations. Addressed deprecation warnings from sklearn in tests and examples. Updated dependency versions due to these changes. Fixed a typo in CONTRIBUTING.rst
