+++ pyproject.toml
@@ -11,10 +11,10 @@ repository = "https://github.com/Sm00thix/IKPLS"
 
 [tool.poetry.dependencies]
 python = ">=3.9, <3.13"
-numpy = "^1.26.1"
-jax = "^0.4.19"
-jaxlib = "^0.4.19"
-scikit-learn = "^1.3.2"
+numpy = "^1.26.3"
+jax = "^0.4.20"
+jaxlib = "^0.4.20"
+scikit-learn = "^1.4.1"
 tqdm = "^4.66.1"
 joblib = "^1.3.2"
 
