### 2024-04-05 #6533 @Sm00thix
> I have a very first and obvious concern about authorship, given that only the corresponding author seems to have contributed to this work and given the nature of the work, which is primarily of an implementation nature, I am questioning the appearance of 3 ! other people in the authors list, and wonder what exactly were their respective contributions ?

Hi @parmentelat
I appreciate your concern.
My co-authors have contributed in writing the paper - even though I am the one who committed it to the repository. They have contributed with discussion and ideas regarding how to communicate the performance of the different algorithms - resulting in the figure in the paper. In general, they have contributed by providing me with guidance throughout the project.
