+++ .github/workflows/pull_request.yml
@@ -5,15 +5,16 @@ name: Python package
 
 on:
   pull_request:
-    branches: [ "main" ]
+    branches: ["main"]
 
 jobs:
   test_package:
-    runs-on: ${{ matrix.os }}  
-    strategy:  
-      fail-fast: false  
-      matrix:  
-        os: [ubuntu-latest, windows-latest, macos-latest] 
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main' && (github.event.paths == null || contains(github.event.paths, 'ikpls/') || contains(github.event.paths, 'pyproject.toml'))
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
     steps:
       - uses: actions/checkout@v4
