+++ .github/workflows/workflow.yml
@@ -6,6 +6,10 @@ name: Python package
 on:
   push:
     branches: [ "main" ]
+    paths:
+      - 'ikpls/**'
+      - 'tests/**'
+      - 'pyproject.toml'
 
 jobs:
   test_package:
