### 2024-04-09 #6533 @Sm00thix
> > > I have a very first and obvious concern about authorship, given that only the corresponding author seems to have contributed to this work and given the nature of the work, which is primarily of an implementation nature, I am questioning the appearance of 3 ! other people in the authors list, and wonder what exactly were their respective contributions ?
> >
> >
> > Hi @parmentelat I appreciate your concern. My co-authors have contributed in writing the paper - even though I am the one who committed it to the repository. They have contributed with discussion and ideas regarding how to communicate the performance of the different algorithms - resulting in the figure in the paper. In general, they have contributed by providing me with guidance throughout the project.
>
> Hi @Sm00thix & @parmentelat,
>
> I have asked my co-editors what the JOSS policy was wrt authorship. And there is actually a paragraph in the journal documentation that mentions that "active project direction and other forms of non-code contributions are considered sufficient for co-authorship of JOSS submissions". (Source: https://joss.readthedocs.io/en/latest/submitting.html#authorship)
>
> So given the details provided by @Sm00thix, I'd say that the list of co-authors is fine! 👍

Hi @boisgera,

Thank you for the clarification! Given this information, I will keep the list of co-authors as is.
