+++ ikpls/fast_cross_validation/numpy_ikpls.py
@@ -12,7 +12,8 @@ E-mail: ole.e@di.ku.dk
 """
 
 import warnings
-from typing import Any, Callable, Union
+from collections import defaultdict
+from typing import Any, Callable, Hashable, Iterable, Union
 
 import joblib
 import numpy as np
@@ -112,7 +113,7 @@ class PLS:
 
     def _stateless_fit(
         self,
-        validation_indices: npt.ArrayLike,
+        validation_indices: list[int],
     ) -> Union[
         tuple[
             npt.NDArray[np.float_],
@@ -143,8 +144,8 @@ class PLS:
 
         Parameters
         ----------
-        validation_indices : Array of shape (N,)
-            Boolean array defining indices into X and Y corresponding to validation
+        validation_indices : List of length N_val
+            Integer list defining indices into X and Y corresponding to validation
             samples.
 
         Returns
@@ -164,7 +165,7 @@ class PLS:
         R : Array of shape (A, K)
             PLS weights matrix to compute scores T directly from original X.
 
-        T : Array of shape (A, N)
+        T : Array of shape (A, N_train)
             PLS scores matrix of X. Only Returned for Improved Kernel PLS Algorithm #1.
 
         training_X_mean : Array of shape (1, K)
@@ -201,17 +202,16 @@ class PLS:
         P = PT.T
         Q = QT.T
         R = RT.T
+
+        validation_size = len(validation_indices)
         if self.algorithm == 1:
-            TT = np.zeros(
-                shape=(self.A, self.N - np.sum(validation_indices)), dtype=self.dtype
-            )
+            TT = np.zeros(shape=(self.A, self.N - validation_size), dtype=self.dtype)
             T = TT.T
 
         # Extract training XTY
         validation_X = self.X[validation_indices]
         validation_Y = self.Y[validation_indices]
         if self.center:
-            validation_size = np.einsum("i->", validation_indices, dtype=int)
             training_size = self.N - validation_size
             N_total_over_N_train = self.N / training_size
             N_val_over_N_train = validation_size / training_size
@@ -269,7 +269,9 @@ class PLS:
                 # Apply the training set scaling
                 training_XTY = training_XTY / (training_X_std.T @ training_Y_std)
         if self.algorithm == 1:
-            training_X = self.X[~validation_indices]
+            training_indices = np.ones(self.N, dtype=bool)
+            training_indices[validation_indices] = False
+            training_X = self.X[training_indices]
             if self.center:
                 # Apply the training set centering
                 training_X = training_X - training_X_mean
@@ -376,7 +378,7 @@ class PLS:
 
     def _stateless_predict(
         self,
-        indices: npt.NDArray[np.float_],
+        indices: list[int],
         B: npt.NDArray[np.float_],
         training_X_mean: npt.NDArray[np.float_],
         training_Y_mean: npt.NDArray[np.float_],
@@ -391,7 +393,7 @@ class PLS:
 
         Parameters
         ----------
-        indices : Array of shape (N,)
+        indices : Array of shape (N_pred,)
             Boolean array defining indices into X and Y corresponding to samples on
             which to predict.
 
@@ -441,7 +443,7 @@ class PLS:
 
     def _stateless_fit_predict_eval(
         self,
-        validation_indices: npt.NDArray[np.float_],
+        validation_indices: list[int],
         metric_function: Callable[
             [npt.NDArray[np.float_], npt.NDArray[np.float_]], Any
         ],
@@ -455,8 +457,8 @@ class PLS:
 
         Parameters
         ----------
-        validation_indices : Array of shape (N,)
-            Boolean array defining indices into X and Y corresponding to validation
+        validation_indices : List of length N_val
+            Integer list defining indices into X and Y corresponding to validation
[…]
