+++ tests/test_ikpls.py
@@ -1432,9 +1432,7 @@ class TestClass:
         sk_x_loadings_sign_flip = np.sign(sk_pls.x_loadings_ / expected_x_loadings)
         sk_x_weights_sign_flip = np.sign(sk_pls.x_weights_ / expected_x_weights)
         sk_y_loadings_sign_flip = np.sign(sk_pls.y_loadings_ / expected_y_loadings)
-        assert_allclose(
-            sk_x_loadings_sign_flip, sk_x_weights_sign_flip, atol=0, rtol=0
-        )
+        assert_allclose(sk_x_loadings_sign_flip, sk_x_weights_sign_flip, atol=0, rtol=0)
         assert_allclose(
             sk_x_loadings_sign_flip, sk_y_loadings_sign_flip, atol=0, rtol=0
         )
@@ -1446,10 +1444,7 @@ class TestClass:
             np_alg_1_x_loadings_sign_flip, np_alg_1_x_weights_sign_flip, atol=0, rtol=0
         )
         assert_allclose(
-            np_alg_1_x_loadings_sign_flip,
-            np_alg_1_y_loadings_sign_flip,
-            atol=0,
-            rtol=0
+            np_alg_1_x_loadings_sign_flip, np_alg_1_y_loadings_sign_flip, atol=0, rtol=0
         )
 
         np_alg_2_x_loadings_sign_flip = np.sign(np_pls_alg_2.P / expected_x_loadings)
@@ -1459,10 +1454,7 @@ class TestClass:
             np_alg_2_x_loadings_sign_flip, np_alg_2_x_weights_sign_flip, atol=0, rtol=0
         )
         assert_allclose(
-            np_alg_2_x_loadings_sign_flip,
-            np_alg_2_y_loadings_sign_flip,
-            atol=0,
-            rtol=0
+            np_alg_2_x_loadings_sign_flip, np_alg_2_y_loadings_sign_flip, atol=0, rtol=0
         )
 
         jax_alg_1_x_loadings_sign_flip = np.sign(jax_pls_alg_1.P / expected_x_loadings)
@@ -1614,16 +1606,10 @@ class TestClass:
             np.abs(jax_pls_alg_2.P), np.abs(expected_x_loadings), atol=3e-6, rtol=0
         )
         assert_allclose(
-            np.abs(diff_jax_pls_alg_1.P),
-            np.abs(expected_x_loadings),
-            atol=3e-6,
-            rtol=0
+            np.abs(diff_jax_pls_alg_1.P), np.abs(expected_x_loadings), atol=3e-6, rtol=0
         )
         assert_allclose(
-            np.abs(diff_jax_pls_alg_2.P),
-            np.abs(expected_x_loadings),
-            atol=3e-6,
-            rtol=0
+            np.abs(diff_jax_pls_alg_2.P), np.abs(expected_x_loadings), atol=3e-6, rtol=0
         )
 
         # Check for expected Y loadings
@@ -1643,16 +1629,10 @@ class TestClass:
             np.abs(jax_pls_alg_2.Q), np.abs(expected_y_loadings), atol=3e-6, rtol=0
         )
         assert_allclose(
-            np.abs(diff_jax_pls_alg_1.Q),
-            np.abs(expected_y_loadings),
-            atol=3e-6,
-            rtol=0
+            np.abs(diff_jax_pls_alg_1.Q), np.abs(expected_y_loadings), atol=3e-6, rtol=0
         )
         assert_allclose(
-            np.abs(diff_jax_pls_alg_2.Q),
-            np.abs(expected_y_loadings),
-            atol=3e-6,
-            rtol=0
+            np.abs(diff_jax_pls_alg_2.Q), np.abs(expected_y_loadings), atol=3e-6, rtol=0
[…]
