### 2024-04-09 Sm00thix/IKPLS#15 @Sm00thix
Selection of validation indices in fast cross-validation algorithm
Hi @parmentelat,

I have realized that there is a subtle error in my implementation of the fast cross-validation algorithm.
In practice, the error has close to no effect but it does change the theoretical runtime of the algorithms explained in https://arxiv.org/abs/2401.13185 from Theta(NK(K+M)) to Theta(NK(K+M) + NP) where P is the number of cross-validation splits.

The error arises due to the selection of validation-indices which, for each cross-validation split, I construct a boolean mask of length N specifying the validation set. This takes NP time.
I have implemented a solution that does not construct a boolean mask of length N for each split. Instead, the new solution constructs for each validation split a list of length N_val, specifying indices into X and Y corresponding to the validation set for the current split. This restores the asymptotic runtime to Theta(NK(K+M)).

In practice this change does not really have an impact on the running due to the practical constants hidden by the Theta-notation. As a practical test, I re-ran the most extreme example in the benchmarks, that is
`time_pls.py -model fastnp2 -n 1000000 -k 500 -m 10 -n_components 30 -n_splits 1000000 -n_jobs -1`
and found that the change in runtime is so small that it is not even visible on the graph.

Additionally, a side-effect of my proposed solution is that the argument `cv_splits` to  `ikpls.fast_cross_validation.numpy_ikpls.PLS.cross_validate` can now be any iterable of hashable (due to the internal usage of dictionaries) instead of just integer arrays. This allows a user to define their cross-validation splits using, e.g., strings if they like. However, it no longer sorts the validation splits. Instead, the resulting metrics will be sorted according to the first occurence of each unique validation split in `cv_splits`.

I have updated the tests and doc strings to reflect these changes.

I do apologize for any disruption of your work in relation to the review at https://github.com/openjournals/joss-reviews/issues/6533

My questions are:
1) Will you accept these changes to the code?
2) Will you accept the timings.png as is or do you require that I rerun the benchmarks related to the fast cross-validation algorithms?

The changes are reflected in https://github.com/Sm00thix/IKPLS/commit/9efdc1af22049cf6065b15bdf280b1ad761895cb
