### 2024-04-09 Sm00thix/IKPLS#15 @Sm00thix
I also wonder if adding the option to return the order, in which cross-validation was performed, would improve usability.
