### 2024-04-09 d4b16a6
Updated docstring and computation of training idxs for algorithm 1 in fast cross-validation.
