### 2024-04-09 697cc48
Updated fast cross-validation alg 1 to use NumPys setdiff instead of built-in Python sets.
