### 2024-04-13 8dff883
Fixed issue with using closure
