### 2024-04-13 Sm00thix/IKPLS#15 @Sm00thix
Hi again @parmentelat,
In relation to my previous comment in this thread, I just realized that I had been running the timing experiments with the original ikpls package instead of the updated one. Indeed, executing `time_pls.py` on the updated code would crash due to some issues with timeit and closures. I have fixed this issue in https://github.com/Sm00thix/IKPLS/commit/8dff883504759f024ae3b0f2d3925526511357e7
I just reran `time_pls.py -model fastnp2 -n 1000000 -k 500 -m 10 -n_components 30 -n_splits 1000000 -n_jobs -1` with the updated code. And my changes seem to have improved the runtime from 19.3 minutes to approximately 17 minutes. This is, however, as argued above, the most extreme example. Given the logarithmic scale of the graphs, this is unlikely to visually change anything. However, as previously mentioned, I will rerun the experiments using the fast cross-validaiton algorithms, if you think that is better.

I am very sorry for any potential confusion caused by this thread.
