### 2024-04-13 ab12af1
Fixed docstrings and made array casting explicit.
