### 2024-04-18 Sm00thix/IKPLS#9 @Sm00thix
I have merged these changes to the main branch. Do you agree that the issue is resolved? @parmentelat
After my previous comment, some additional changes were made. Now, executing pylint with --max-line-length 88 --disable invalid-name will now give scores of
8.81/10 for examples/
9.24/10 for ikpls/
9.09/10 for tests/
