### 2024-04-20 Sm00thix/IKPLS#15 @parmentelat
> I also wonder if adding the option to return the order, in which cross-validation was performed, would improve usability. Do you have any comment to this?

in general I'd tend to say yes; can you be more specific on how you'd propose to g about doing this ?
