### 2024-04-20 Sm00thix/IKPLS#15 @parmentelat
and oops I forgot to link the meta-issue
https://github.com/openjournals/joss-reviews/issues/6533
