### 2024-04-20 #6533 @parmentelat
> So given the details provided by @Sm00thix, I'd say that the list of co-authors is fine! 👍

I abide by the decision of the editor, however I can't help but think that as phrased above, this rationale can apply to any contribution in any environment, and personally I would have appreciated a little more explicit list of contributions !
