### 2024-04-21 Sm00thix/IKPLS#15 @Sm00thix
> > My questions are:
>
> 1. Will you accept these changes to the code?
> 2. Will you accept the timings.png as is or do you require that I rerun the benchmarks related to the fast cross-validation algorithms?
>
> my answers
>
> 1. yes
> 2. it's best if you can make sure to provide a figure that reflects a specific version of the code - and to mention that version in the paper itself

Thank you for the feedback. I will make sure to do as you suggest in 2.
