### 2024-04-21 Sm00thix/IKPLS#15 @Sm00thix
> > I also wonder if adding the option to return the order, in which cross-validation was performed, would improve usability. Do you have any comment to this?
>
> in general I'd tend to say yes; can you be more specific on how you'd propose to g about doing this ?

Currently the cross_validate method returns a list[Any] of length P, where P is the number of cross_validation splits as determined by the number unique elements passed in the cv_splits:Iterable[Hashable] argument to cross_validate. This list contains the result of evaluating metric_function on each validation split.
I was thinking along the lines of having cross_validate returning dict[Any] containing as keys the P unique elements in cv_splits and as values the results of evaluating metric_function on each of the P validation splits. Then a user can directly use the name of a validation split to lookup the results of that split in the returned dict. And the order of dict.keys() will be the same as the order in which validation splits were visited. Does my explanation make sense? Otherwise, I can implement it, and we can roll it back if you do not like it.
