### 2024-04-22 840fb3a
Updated CONTRIBUTING.rst with more visible guides for testing
