### 2024-04-22 75445c0
Fixed cross-reference in CONTRIBUTING.rst
