### 2024-04-22 e45a210
Made a copyable codeblock for the test command in CONTRIBUTING.rst
