### 2024-04-22 1bdd9f9
Expanding upon test suite in CONTRIBUTING.rst
