### 2024-04-22 5d68c50
Added instructions on how to build docs.
