+++ docs/source/fast_cross_validation/numpy_ikpls.rst
@@ -1,6 +1,6 @@
-^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 ikpls.fast\_cross\_validation.numpy\_ikpls
-^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. automodule:: ikpls.fast_cross_validation.numpy_ikpls
    
