### 2024-04-22 f194deb
Added benchmarking section in CONTRIBUTING.rst
