### 2024-04-22 f5d16fe
include contributing guidelines as the last section in sphinx doc
