### 2024-04-23 a051a69
Fixed enumeration in CONTRIBUTING.rst
