### 2024-04-23 78072b7
Gave docs/CONTRIBUTING.rst a header
