### 2024-04-23 0e52fb1
Attempting to fix readthedocs for CONTRIBUTING.md
