### 2024-04-23 ba988a3
Updated documentation to use CONTRIBUTING.md
