+++ docs/CONTRIBUTING.rst
@@ -1,4 +1,0 @@
-Contribution Guidelines
-=======================
-
-.. include:: CONTRIBUTING.md
