### 2024-04-23 ecd525a
Modified CONTRIBUTING.md to work with sphinx
