+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 ikpls
 sphinx-rtd-theme
+myst-parser
