### 2024-04-23 476ca02
Changed README from rst to md and updated docs accordingly.
