+++ docs/index.rst
@@ -1,5 +1,7 @@
-.. include:: ../README.rst
-
+.. include:: README.md
+    :parser: myst_parser.sphinx_
+    :start-after: <!-- index.rst content start -->
+    
 .. toctree::
     :caption: Table of Contents
     :titlesonly:
