+++ docs/README.md
@@ -1,0 +1,3 @@
+```{include} ../README.md
+```
+<!-- index.rst content start -->
