+++ docs/CONTRIBUTING.md
@@ -1,1 +1,2 @@
 ```{include} ../CONTRIBUTING.md
+```
