### 2024-04-23 80cb423
Updated outdated CONTRIBUTING.rst link to CONTRIBUTING.md in README.md
