+++ README.rst
@@ -1,157 +1,0 @@
-Improved Kernel Partial Least Squares (IKPLS) and Fast Cross-Validation
-=======================================================================
-
-.. image:: https://img.shields.io/pypi/v/ikpls.svg
-   :target: https://pypi.python.org/pypi/ikpls/
-   :alt: PyPI Version
-.. image:: https://img.shields.io/pypi/dm/ikpls
-   :target: https://pypi.python.org/pypi/ikpls/
-   :alt: PyPI - Downloads
-.. image:: https://img.shields.io/pypi/pyversions/ikpls.svg
-   :target: https://pypi.python.org/pypi/ikpls/
-   :alt: Python Versions
-.. image:: https://img.shields.io/pypi/l/ikpls.svg
-   :target: https://pypi.python.org/pypi/ikpls/
-   :alt: License
-.. image:: https://readthedocs.org/projects/ikpls/badge/?version=latest
-   :target: https://ikpls.readthedocs.io/en/latest/?badge=latest
-   :alt: Documentation Status
-.. image:: https://github.com/Sm00thix/IKPLS/actions/workflows/workflow.yml/badge.svg
-   :target: https://github.com/Sm00thix/IKPLS/actions/workflows/workflow.yml
-   :alt: Build Status
-.. image:: https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc/status.svg
-   :target: https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc
-   :alt: JOSS Status
-
-Fast CPU, GPU, and TPU Python implementations of Improved Kernel PLS Algorithm #1 and Algorithm #2 by Dayal and MacGregor [1]_. Improved Kernel PLS is both fast [2]_ and numerically stable [3]_.
-The CPU implementations use NumPy [4]_ and subclass BaseEstimator from scikit-learn [5]_, allowing integration into scikit-learn's ecosystem of machine learning algorithms and pipelines. For example, the CPU implementations can be used with scikit-learn's `cross_validate <https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_validate.html>`_.
-The GPU and TPU implementations use Google's JAX [6]_. JAX supports automatic differentiation while allowing CPU, GPU, and TPU execution. This implies that the JAX implementations can be combined with deep learning approaches, as the PLS fit is differentiable.
-
-The documentation is available at https://ikpls.readthedocs.io/en/latest/; examples can be found at https://github.com/Sm00thix/IKPLS/tree/main/examples.
-
-.. [1] `Dayal, B. S., & MacGregor, J. F. (1997). Improved PLS algorithms. Journal of Chemometrics: A Journal of the Chemometrics Society, 11(1), 73-85`_.
-.. [2] `Alin, A. (2009). Comparison of PLS algorithms when the number of objects is much larger than the number of variables. Statistical papers, 50, 711-720`_.
-.. [3] `Andersson, M. (2009). A comparison of nine PLS1 algorithms. Journal of Chemometrics: A Journal of the Chemometrics Society, 23(10), 518-529`_.
-.. [4] `NumPy`_.
-.. [5] `scikit-learn`_.
-.. [6] `JAX`_.
-
-.. _Dayal, B. S., & MacGregor, J. F. (1997). Improved PLS algorithms. Journal of Chemometrics\: A Journal of the Chemometrics Society, 11(1), 73-85: https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23?
[…]
