+++ README.md
@@ -1,0 +1,181 @@
+# Improved Kernel Partial Least Squares (IKPLS) and Fast Cross-Validation
+
+[![PyPI Version](https://img.shields.io/pypi/v/ikpls.svg)](https://pypi.python.org/pypi/ikpls/)
+
+[![PyPI - Downloads](https://img.shields.io/pypi/dm/ikpls)](https://pypi.python.org/pypi/ikpls/)
+
+[![Python Versions](https://img.shields.io/pypi/pyversions/ikpls.svg)](https://pypi.python.org/pypi/ikpls/)
+
+[![License](https://img.shields.io/pypi/l/ikpls.svg)](https://pypi.python.org/pypi/ikpls/)
+
+[![Documentation Status](https://readthedocs.org/projects/ikpls/badge/?version=latest)](https://ikpls.readthedocs.io/en/latest/?badge=latest)
+
+[![Build Status](https://github.com/Sm00thix/IKPLS/actions/workflows/workflow.yml/badge.svg)](https://github.com/Sm00thix/IKPLS/actions/workflows/workflow.yml)
+
+[![JOSS Status](https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc/status.svg)](https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc)
+
+Fast CPU, GPU, and TPU Python implementations of Improved Kernel PLS
+Algorithm #1 and Algorithm #2 by Dayal and MacGregor[^1]. Improved
+Kernel PLS is both fast[^2] and numerically stable[^3]. The CPU
+implementations use NumPy[^4] and subclass BaseEstimator from
+scikit-learn[^5], allowing integration into scikit-learn\'s ecosystem of
+machine learning algorithms and pipelines. For example, the CPU
+implementations can be used with scikit-learn\'s
+[cross_validate](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_validate.html).
+The GPU and TPU implementations use Google\'s JAX[^6]. JAX supports
+automatic differentiation while allowing CPU, GPU, and TPU execution.
+This implies that the JAX implementations can be combined with deep
+learning approaches, as the PLS fit is differentiable.
+
+The documentation is available at
+<https://ikpls.readthedocs.io/en/latest/>; examples can be found at
+<https://github.com/Sm00thix/IKPLS/tree/main/examples>.
+
+## Fast Cross-Validation
+
+In addition to the implementations mentioned above, this package
+contains the novel, fast cross-validation algorithms mentioned in[^7]
+using both IKPLS algorithms. The fast cross-validation algorithms
+benefit both IKPLS Algorithms and especially Algorithm #2. The fast
+cross-validation algorithms are mathematically equivalent to the
+classical cross-validation algorithm. Still, they are much quicker if
+cross-validation splits exceed 3. The fast cross-validation algorithms
+correctly handle (column-wise) centering and scaling of the X and Y
+input matrices using training set means and standard deviations to avoid
+data leakage from the validation set. This centering and scaling can be
+enabled by setting the center parameter to True and the scale parameter
+to True, respectively. The fast cross-validation algorithms correctly
+handle row-wise preprocessing such as (row-wise) centering and scaling
+of the X and Y input matrices, convolution, or other preprocessing.
+Row-wise preprocessing can safely be applied before passing the data to
[…]
