+++ docs/README.md
@@ -1,3 +1,2 @@
 ```{include} ../README.md
 ```
-<!-- index.rst content start -->
