+++ docs/index.rst
@@ -1,6 +1,5 @@
 .. include:: README.md
     :parser: myst_parser.sphinx_
-    :start-after: <!-- index.rst content start -->
     
 .. toctree::
     :caption: Table of Contents
