+++ README.md
@@ -15,14 +15,14 @@
 [![JOSS Status](https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc/status.svg)](https://joss.theoj.org/papers/ac559cbcdc6e6551f58bb3e573a73afc)
 
 Fast CPU, GPU, and TPU Python implementations of Improved Kernel PLS
-Algorithm #1 and Algorithm #2 by Dayal and MacGregor[^1]. Improved
-Kernel PLS is both fast[^2] and numerically stable[^3]. The CPU
-implementations use NumPy[^4] and subclass BaseEstimator from
-scikit-learn[^5], allowing integration into scikit-learn\'s ecosystem of
+Algorithm #1 and Algorithm #2 ([1]). Improved
+Kernel PLS is both fast ([2]) and numerically stable ([3]). The CPU
+implementations use ([4]) and subclass BaseEstimator from scikit-learn ([5]),
+allowing integration into scikit-learn\'s ecosystem of
 machine learning algorithms and pipelines. For example, the CPU
 implementations can be used with scikit-learn\'s
 [cross_validate](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_validate.html).
-The GPU and TPU implementations use Google\'s JAX[^6]. JAX supports
+The GPU and TPU implementations use Google\'s JAX ([6]). JAX supports
 automatic differentiation while allowing CPU, GPU, and TPU execution.
 This implies that the JAX implementations can be combined with deep
 learning approaches, as the PLS fit is differentiable.
@@ -34,7 +34,7 @@ The documentation is available at
 ## Fast Cross-Validation
 
 In addition to the implementations mentioned above, this package
-contains the novel, fast cross-validation algorithms mentioned in[^7]
+contains the novel, fast cross-validation algorithms by Engstrøm 
 using both IKPLS algorithms. The fast cross-validation algorithms
 benefit both IKPLS Algorithms and especially Algorithm #2. The fast
 cross-validation algorithms are mathematically equivalent to the
@@ -155,27 +155,26 @@ will find:
 To contribute, please read the [Contribution
 Guidelines](https://github.com/Sm00thix/IKPLS/blob/main/CONTRIBUTING.md).
 
-[^1]: [Dayal, B. S., & MacGregor, J. F. (1997). Improved PLS algorithms.
-    Journal of Chemometrics: A Journal of the Chemometrics Society,
-    11(1),
-    73-85](https://doi.org/10.1002/(SICI)1099-128X(199701)11:1%3C73::AID-CEM435%3E3.0.CO;2-%23?).
+## References
 
-[^2]: [Alin, A. (2009). Comparison of PLS algorithms when the number of
-    objects is much larger than the number of variables. Statistical
-    papers, 50, 711-720](https://doi.org/10.1007/s00362-009-0251-7).
+## References
 
-[^3]: [Andersson, M. (2009). A comparison of nine PLS1 algorithms.
-    Journal of Chemometrics: A Journal of the Chemometrics Society,
-    23(10), 518-529](https://doi.org/10.1002/cem.1248).
-
-[^4]: [NumPy](https://numpy.org/).
-
-[^5]: [scikit-learn](https://scikit-learn.org/stable/).
-
-[^6]: [JAX](https://jax.readthedocs.io/en/latest/).
-
-[^7]: [Engstrøm, O.-C. G. (2024). Shortcutting Cross-Validation:
[…]
