+++ CONTRIBUTING.md
@@ -30,19 +30,18 @@ The sections below outline the steps in each case.
 1. (**Important**) Announce your plan to the rest of the community *before you start working*. This announcement should be in the form of a (new) issue;
 2. (**Important**) Wait until some kind of consensus is reached about your idea being a good idea;
 3. If needed, fork the repository to your own Github profile and create your own feature branch off of the latest main commit. While working on your feature branch, make sure to stay up to date with the main branch by pulling in changes, possibly from the 'upstream' repository (follow the instructions [here](https://help.github.com/articles/configuring-a-remote-for-a-fork/) and [here](https://help.github.com/articles/syncing-a-fork/));
-4. Make sure the existing tests still work by following the instructions in [Run the test suite](#testing);
+4. Make sure the existing tests still work by following the instructions in [Run the test suite](#run-the-test-suite);
 5. Add your own tests (if necessary);
-6. If you added your own tests, make sure they pass by following the instructions in [Run the test suite](#testing);
+6. If you added your own tests, make sure they pass by following the instructions in [Run the test suite](#run-the-test-suite);
 7. If your contribution is a performance enhancement, make sure to include benchmarks. See [Benchmarking](#benchmarking) for more information;
 8. Update or expand the documentation;
-9. Make sure the documentation builds without errors by following the instructions in [Build the documentation](#documentation);
+9. Make sure the documentation builds without errors by following the instructions in [Build the documentation](#build-the-documentation);
 10. Check that you can build the package locally and that it passes twine check. See [Build from source](#build-from-source) for more information;
 11. [Push](http://rogerdudler.github.io/git-guide/) your feature branch to (your fork of) the IKPLS repository on GitHub;
 12. Create the pull request, e.g. following the instructions [here](https://help.github.com/articles/creating-a-pull-request/).
 
 If you feel like you've made a valuable contribution, but you don't know how to write or run tests for it or generate the documentation, don't let this discourage you from making the pull request; we can help you! Just submit the pull request, but remember that you might be asked to append additional commits to your pull request.
 
-(build-from-source)=
 ## Build from source
 
 IKPLS uses [poetry](https://python-poetry.org/) to manage its dependencies and packaging. To build the package from source, follow these steps:
@@ -84,7 +83,6 @@ IKPLS uses [poetry](https://python-poetry.org/) to manage its dependencies and p
     twine check dist/*
     ```
 
-(testing)=
 ## Run the test suite
 
 To run the test suite, follow these steps:
@@ -113,7 +111,6 @@ To run the test suite, follow these steps:
     poetry run pytest tests --doctest-modules --junitxml=junit/test-results.xml --cov=ikpls/ --cov-report=xml --cov-report=html --typeguard-packages=ikpls/
     ```
 
-(documentation)=
 ## Build the documentation
 
 1. Install sphinx and the sphinx-rtd-theme:
@@ -134,7 +131,6 @@ To run the test suite, follow these steps:
     make html
     ```
 
-(benchmarking)=
 ## Benchmarking
 
 To run benchmarks, follow the instructions [here](https://github.com/Sm00thix/IKPLS/blob/main/paper/README.md).
