### 2024-04-24 1650c10
Updated pyproject.toml to use README.md instead of README.rst. Also updated version number.
