+++ ikpls/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.0.post1"
+__version__ = "1.2.0.post2"
