### 2024-04-24 ea8a206
Updated fast cv algorithms to use dicts and updated benchmarks to show ikpls version. This is related to #15
