### 2024-04-25 239e28f
In README, removed MyST syntax: (references)=. This is related to #19
