### 2024-04-25 4eade46
Updated test action to explicitly set arch to x64 given macos and Python3.9. Arm is no longer supported for this combination
