+++ .github/workflows/workflow.yml
@@ -24,6 +24,7 @@ jobs:
       - uses: ./.github/actions/test
         with:
           PYTHON_VERSION: ${{ matrix.python-version }}
+          OS: ${{ matrix.os }}
 
   build_package:
     runs-on: ubuntu-latest
