### 2024-04-25 b6b9070
Updated workflow to also trigger on changes to .github/
