### 2024-04-25 42b8d92
Updated workflow to also trigger on changes to .github/
