### 2024-04-25 4254a8e
Updated workflow file to attempt to fix action issue with architecture for macos with python3.9
