### 2024-04-25 901d429
Updated workflow file to attempt to fix action issue with architecture for macos with python3.9
