+++ .github/actions/test/action.yml
@@ -5,19 +5,14 @@ inputs:
   PYTHON_VERSION:
     description: Test Python version
     required: true
-  OS:
-    description: Test OS
-    required: true
 
 runs:
   using: "composite"
   steps:
     - name: Set up Python
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ inputs.PYTHON_VERSION }}
-        # Conditionally set architecture
-        architecture: ${{ 'arm64' if inputs.OS == 'macos-latest' and int(inputs.PYTHON_VERSION.split('.')[1]) > 9 else 'x64' }}
 
     - name: Install Poetry
       run: |
