### 2024-04-25 325d71e
Updated workflow file to attempt to fix action issue with architecture for macos with python3.9
