### 2024-04-25 443bd16
Exclude the combination of macos-latest and Python3.9 from the testing suite as GitHub action's combination of these is faulty
