+++ .github/actions/test/action.yml
@@ -5,9 +5,6 @@ inputs:
   PYTHON_VERSION:
     description: Test Python version
     required: true
-  OS:
-    description: Test OS
-    required: true
 
 runs:
   using: "composite"
@@ -15,9 +12,7 @@ runs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: ${{ inputs.PYTHON_VERSION }}
-        # If OS is Mac and Python Version == 3.9, use x64, otherwise use default value
-        architecture: ${{ (inputs.OS == 'macos-latest' && inputs.PYTHON_VERSION == '3.9') && 'x64' || '' }}
+        python-version: ${{ inputs.PYTHON_VERSION }
 
     - name: Install Poetry
       run: |
