+++ .github/actions/test/action.yml
@@ -12,7 +12,7 @@ runs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: ${{ inputs.PYTHON_VERSION }
+        python-version: ${{ inputs.PYTHON_VERSION }}
 
     - name: Install Poetry
       run: |
