+++ .github/workflows/workflow.yml
@@ -18,11 +18,8 @@ jobs:
     strategy:  
       fail-fast: false  
       matrix:  
-        os: [ubuntu-latest, windows-latest, macos-latest] 
+        os: [ubuntu-latest, windows-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-        exclude:
-          - os: macos-latest
-            python-version: "3.9"
     steps:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/test
