### 2024-04-25 c1a795b
Temporarily set macos to version 12 due to issues with 14.
