### 2024-04-25 144ec50
Fixed typo in docstring for fast cv
