### 2024-04-25 6a34ca8
Fixed another typo in docstring for fast cv
