+++ ikpls/fast_cross_validation/numpy_ikpls.py
@@ -544,9 +544,9 @@ class PLS:
             An iterable defining cross-validation splits. Each unique value in
             `cv_splits` corresponds to a different fold.
 
-        metric_function : Callable `Y_true`, `Y_pred` -> Any
+        metric_function : callable
             A callable receiving arrays `Y_true` of shape (N_val, M) and `Y_pred` of
-            shape (A, N_val, M) and returning Any Computes a metric based on true
+            shape (A, N_val, M) and returning Any. Computes a metric based on true
             values `Y_true` and predicted values `Y_pred`. `Y_pred` contains a
             prediction for all `A` components.
 
