### 2024-04-25 5eebb53
Removed obsolete part of docstring
