### 2024-04-30 0dc8776
Added support for independent centering and scaling of X and  Y
