+++ ikpls/jax_ikpls_alg_2.py
@@ -28,15 +28,27 @@ class PLS(PLSBase):
 
     Parameters
     ----------
-    center : bool, optional, default=True
-        Whether to center the data before fitting. If True, then the mean of the
-        training data is subtracted from the data. If False, then the data is assumed
-        to be already centered.
+    center_X : bool, default=True
+        Whether to center the predictor variables (X) before fitting. If True, then
+        the mean of the training data is subtracted from the predictor variables.
+        If False, then the data is assumed to be already centered.
 
-    scale : bool, optional, default=True
-        Whether to scale the data before fitting. If True, then the data is scaled
-        using Bessel's correction for the unbiased estimate of the sample standard
-        deviation. If False, then the data is assumed to be already scaled.
+    center_Y : bool, default=True
+        Whether to center the response variables (Y) before fitting. If True, then
+        the mean of the training data is subtracted from the response variables. If
+        False, then the data is assumed to be already centered.
+
+    scale_X : bool, default=True
+        Whether to scale the predictor variables (X) before fitting. If True, then
+        the data is scaled using Bessel's correction for the unbiased estimate of
+        the sample standard deviation. If False, then the data is assumed to be
+        already scaled.
+
+    scale_Y : bool, default=True
+        Whether to scale the response variables (Y) before fitting. If True, then
+        the data is scaled using Bessel's correction for the unbiased estimate of
+        the sample standard deviation. If False, then the data is assumed to be
+        already scaled.
 
     copy : bool, optional, default=True
         Whether to copy `X` and `Y` in fit before potentially applying centering and
@@ -62,16 +74,20 @@ class PLS(PLSBase):
 
     def __init__(
         self,
-        center: bool = True,
-        scale: bool = True,
+        center_X: bool = True,
+        center_Y: bool = True,
+        scale_X: bool = True,
+        scale_Y: bool = True,
         copy: bool = True,
         dtype: DTypeLike = jnp.float64,
         reverse_differentiable: bool = False,
         verbose: bool = False,
     ) -> None:
         super().__init__(
-            center=center,
-            scale=scale,
+            center_X=center_X,
+            center_Y=center_Y,
+            scale_X=scale_X,
+            scale_Y=scale_Y,
             copy=copy,
             dtype=dtype,
             reverse_differentiable=reverse_differentiable,
@@ -323,21 +339,32 @@ class PLS(PLSBase):
         instead of storing them in the class instance.
         """
         self.B, W, P, Q, R, self.X_mean, self.Y_mean, self.X_std, self.Y_std = (
-            self.stateless_fit(X, Y, A, self.center, self.scale, self.copy)
+            self.stateless_fit(
+                X,
+                Y,
+                A,
+                self.center_X,
+                self.center_Y,
+                self.scale_X,
+                self.scale_Y,
+                self.copy,
+            )
         )
         self.W = W.T
         self.P = P.T
         self.Q = Q.T
         self.R = R.T
 
-    @partial(jax.jit, static_argnums=(0, 3, 4, 5, 6))
+    @partial(jax.jit, static_argnums=(0, 3, 4, 5, 6, 7, 8))
     def stateless_fit(
         self,
         X: ArrayLike,
         Y: ArrayLike,
         A: int,
-        center: bool = True,
-        scale: bool = True,
+        center_X: bool = True,
+        center_Y: bool = True,
+        scale_X: bool = True,
+        scale_Y: bool = True,
         copy: bool = True,
     ) -> Tuple[jax.Array, jax.Array, jax.Array, jax.Array, jax.Array]:
         """
@@ -356,6 +383,28 @@ class PLS(PLSBase):
 
         A : int
             Number of components in the PLS model.
+
+        center_X : bool, default=True
+            Whether to center the predictor variables (X) before fitting. If True, then
[…]
