### 2024-04-30 92ce90c
Updated docs. Also reduced number of X features in center scale combination tests in the interest of execution time.
