+++ ikpls/jax_ikpls_alg_2.py
@@ -29,26 +29,22 @@ class PLS(PLSBase):
     Parameters
     ----------
     center_X : bool, default=True
-        Whether to center the predictor variables (X) before fitting. If True, then
-        the mean of the training data is subtracted from the predictor variables.
-        If False, then the data is assumed to be already centered.
-
+        Whether to center `X` before fitting by subtracting its row of
+        column-wise means from each row.
+    
     center_Y : bool, default=True
-        Whether to center the response variables (Y) before fitting. If True, then
-        the mean of the training data is subtracted from the response variables. If
-        False, then the data is assumed to be already centered.
-
+        Whether to center `Y` before fitting by subtracting its row of
+        column-wise means from each row.
+    
     scale_X : bool, default=True
-        Whether to scale the predictor variables (X) before fitting. If True, then
-        the data is scaled using Bessel's correction for the unbiased estimate of
-        the sample standard deviation. If False, then the data is assumed to be
-        already scaled.
+        Whether to scale `X` before fitting by dividing each row with the row of `X`'s
+        column-wise standard deviations. Bessel's correction for the unbiased estimate
+        of the sample standard deviation is used.
 
     scale_Y : bool, default=True
-        Whether to scale the response variables (Y) before fitting. If True, then
-        the data is scaled using Bessel's correction for the unbiased estimate of
-        the sample standard deviation. If False, then the data is assumed to be
-        already scaled.
+        Whether to scale `Y` before fitting by dividing each row with the row of `X`'s
+        column-wise standard deviations. Bessel's correction for the unbiased estimate
+        of the sample standard deviation is used.
 
     copy : bool, optional, default=True
         Whether to copy `X` and `Y` in fit before potentially applying centering and
@@ -70,6 +66,12 @@ class PLS(PLSBase):
         If True, each sub-function will print when it will be JIT compiled. This can be
         useful to track if recompilation is triggered due to passing inputs with
         different shapes.
+
+    Notes
+    -----
+    Any centering and scaling is undone before returning predictions with `fit` to
+    ensure that predictions are on the original scale. If both centering and scaling
+    are True, then the data is first centered and then scaled.
     """
 
     def __init__(
@@ -385,26 +387,22 @@ class PLS(PLSBase):
             Number of components in the PLS model.
 
         center_X : bool, default=True
-            Whether to center the predictor variables (X) before fitting. If True, then
-            the mean of the training data is subtracted from the predictor variables.
-            If False, then the data is assumed to be already centered.
-
+            Whether to center `X` before fitting by subtracting its row of
+            column-wise means from each row.
+        
         center_Y : bool, default=True
-            Whether to center the response variables (Y) before fitting. If True, then
-            the mean of the training data is subtracted from the response variables. If
-            False, then the data is assumed to be already centered.
-
+            Whether to center `Y` before fitting by subtracting its row of
+            column-wise means from each row.
+        
         scale_X : bool, default=True
-            Whether to scale the predictor variables (X) before fitting. If True, then
-            the data is scaled using Bessel's correction for the unbiased estimate of
-            the sample standard deviation. If False, then the data is assumed to be
-            already scaled.
+            Whether to scale `X` before fitting by dividing each row with the row of
+            `X`'s column-wise standard deviations. Bessel's correction for the unbiased
+            estimate of the sample standard deviation is used.
 
         scale_Y : bool, default=True
-            Whether to scale the response variables (Y) before fitting. If True, then
-            the data is scaled using Bessel's correction for the unbiased estimate of
-            the sample standard deviation. If False, then the data is assumed to be
-            already scaled.
+            Whether to scale `Y` before fitting by dividing each row with the row of
+            `X`'s column-wise standard deviations. Bessel's correction for the unbiased
+            estimate of the sample standard deviation is used.
 
         Returns
         -------
