### 2024-04-30 e5a70a0
Fix logical error in one test
Previous update set K=10 and M=10 in test_center_scale_pls_2_m_less_k. I updated it to K=11.
