### 2024-05-05 4201b5e
Added support for independent centering and scaling of X and  Y (#20)
* Added support for independent centering and scaling of X and  Y

* Updated docs.
