+++ ikpls/fast_cross_validation/numpy_ikpls.py
@@ -30,21 +30,27 @@ class PLS:
 
     Parameters
     ----------
-    center : bool, optional default=False
-            Whether to center `X` and `Y` before fitting by subtracting a mean row from
-            each. The centering is computed on the training set for each fold to avoid
-            data leakage. The centering is undone before returning predictions. Setting
-            this to True while using multiple jobs will increase the memory consumption
-            as each job will then have to keep its own copy of the data with its
-            specific centering.
-
-    scale : bool, optional default=False, only used if `center` is True
-        Whether to scale `X` and `Y` before fitting by dividing each row by its
-        standard deviation. The scaling is computed on the training set for each fold
-        to avoid data leakage. The scaling is undone before returning predictions.
-        Setting this to True while using multiple jobs will increase the memory
-        consumption as each job will then have to keep its own copy of the data with
-        its specific scaling.
+    center_X : bool, optional default=True
+        Whether to center `X` before fitting by subtracting its row of
+        column-wise means from each row. The row of column-wise means is computed on
+        the training set for each fold to avoid data leakage.
+    
+    center_Y : bool, optional default=True
+        Whether to center `Y` before fitting by subtracting its row of
+        column-wise means from each row. The row of column-wise means is computed on
+        the training set for each fold to avoid data leakage.
+    
+    scale_X : bool, optional default=True
+        Whether to scale `X` before fitting by dividing each row with the row of `X`'s
+        column-wise standard deviations. Bessel's correction for the unbiased estimate
+        of the sample standard deviation is used. The row of column-wise standard
+        deviations is computed on the training set for each fold to avoid data leakage.
+    
+    scale_Y : bool, optional default=True
+        Whether to scale `Y` before fitting by dividing each row with the row of `X`'s
+        column-wise standard deviations. Bessel's correction for the unbiased estimate
+        of the sample standard deviation is used. The row of column-wise standard
+        deviations is computed on the training set for each fold to avoid data leakage.
 
     algorithm : int, default=1
         Whether to use Improved Kernel PLS Algorithm #1 or #2.
@@ -58,17 +64,32 @@ class PLS:
     ------
     ValueError
         If `algorithm` is not 1 or 2.
+    
+    Notes
+    -----
+    Any centering and scaling is undone before returning predictions to ensure that
+    predictions are on the original scale. If both centering and scaling are True, then
+    the data is first centered and then scaled.
+
+    Setting either of `center_X`, `center_Y`, `scale_X`, or `scale_Y` to True, while
+    using multiple jobs, will increase the memory consumption as each job will then
+    have to keep its own copy of :math:`\mathbf{X}^{\mathbf{T}}\mathbf{X}` and
+    :math:`\mathbf{X}^{\mathbf{T}}\mathbf{Y}` with its specific centering and scaling.
     """
 
     def __init__(
         self,
-        center: bool = True,
-        scale: bool = True,
+        center_X: bool = True,
+        center_Y: bool = True,
+        scale_X: bool = True,
+        scale_Y: bool = True,
         algorithm: int = 1,
         dtype: np.float_ = np.float64,
     ) -> None:
-        self.center = center
-        self.scale = scale
+        self.center_X = center_X
+        self.center_Y = center_Y
+        self.scale_X = scale_X
+        self.scale_Y = scale_Y
         self.algorithm = algorithm
         self.dtype = dtype
         self.name = f"Improved Kernel PLS Algorithm #{algorithm}"
@@ -209,90 +230,113 @@ class PLS:
             TT = np.zeros(shape=(self.A, self.N - validation_size), dtype=self.dtype)
             T = TT.T
 
-        # Extract training XTY
+        # Extract training X and Y
[…]
