### 2024-05-09 #6533 @boisgera
Hi @Sm00thix,

To let you know what's going on:

  - I have exchanged some e-mails with @parmentelat wrt the review progress (basically: 👍).

  - @basileMarchand is not currently offline, but I'll try to have an exchange with him next week.
