### 2024-05-24 Sm00thix/IKPLS#21 @Sm00thix
License appendix
Hi @parmentelat ,

You previously helped me with the LICENSE in relation to your review over at [JOSS](https://github.com/openjournals/joss-reviews/issues/6533). I'm wondering if the following part is meant to be deleted given that I have already followed the instructions in that part. Will you clarify whether or not I should delete this part?

Best,
Ole

>    APPENDIX: How to apply the Apache License to your work.
>
>       To apply the Apache License to your work, attach the following
>       boilerplate notice, with the fields enclosed by brackets "[]"
>       replaced with your own identifying information. (Don't include
>       the brackets!)  The text should be enclosed in the appropriate
>       comment syntax for the file format. We also recommend that a
>       file or class name and description of purpose be included on the
>       same "printed page" as the copyright notice for easier
>       identification within third-party archives.
