### 2024-05-27 bfae197
Update LICENSE in relation to #21
