### 2024-05-30 9c354dd
Add instruction to install MyST-Parser in CONTRIBUTING.md
MyST-Parser is required to build the documentation. This commit adds an instruction to install the parser when building the documentation.
