### 2024-06-06 7af091e
Added package description for PyPI
