### 2024-06-19 80c4ca4
improved paper/README.md
outline the different steps i.e.
   one run / multiple runs / produce the picture
document the new notebook
pleased the linter in the mix
