### 2024-06-19 b176fea
Updated scikit-learn dependency to require at least 1.5.0. Related to #22.
