+++ pyproject.toml
@@ -14,7 +14,7 @@ python = ">=3.9, <3.13"
 numpy = "^1.26.3"
 jax = "^0.4.20"
 jaxlib = "^0.4.20"
-scikit-learn = "^1.4.1"
+scikit-learn = "^1.5.0"
 tqdm = "^4.66.1"
 joblib = "^1.3.2"
 
