### 2024-06-19 515ee78
a notebook to reproduce timings.csv (#23)
* can choose output file with time_pls.py --output

* a notebook for reproducing all the results in the picture

* improved paper/README.md

outline the different steps i.e.
   one run / multiple runs / produce the picture
document the new notebook
pleased the linter in the mix
