### 2024-06-19 #6533 @parmentelat
I have now checked all the boxes, pending a few minor changes that will be raised by @basileMarchand with whom we had a meeting together with @boisgera earlier today

here's a couple general comments/reservations that I'd like to make before closing my contribution to his review

### About contribution and authorship:

Contributions to the project, in practical terms and in particular with respect to commits in the repo, have been exclusively made by the submitting author @Sm00thix; the contributions of the 3 other authors mentioned in the paper are far more fuzzy, and have been described in very vague terms as "discussions and feedback on the project", as reported by the main author upon our request for clarification; nor have they taken any part during the paper review.

Although this seems to be in line with the authorship guidelines of the journal, and this has been deemed acceptable by the editor, I must admit that I find this practice to be somewhat questionable, as it is not clear to me how the contributions of the 3 other authors have been substantial enough to warrant their inclusion in the author list.

### About functionality:

For practical reasons I have only been able to run the code on a CPU, and have not been able to verify the GPU / TPU functionality, that is advertised as a side-effect of the JAX implementation. However, the code seems to be well-structured and well-documented, and I have no reason to doubt that the GPU functionality works as intended.
