### 2024-06-19 e1c3922
more changes to allow to choose the name for the .csv and .png result files (#25)
* plot_timings.py accepts the input csv as a --input option

* allow to choose an output csv file from reproducing-results-notebook.py as well
