### 2024-06-19 Sm00thix/IKPLS#26 @basileMarchand
Distinguish between Test and Packaging in GitHub Actions and add separate badges
Hello

For automated tests, IKPLS uses GitHub Actions to automate the testing, packaging, and deployment processes.
However, in the GitHub Actions interface, it is difficult to access on the test results because it is not clearly distinguished from the packaging part and in addition the two workflows are named "Python Package". Additionally, the project's README has a packaging badge but no Test badge.

Could you please:

- Clearly distinguish between test and packaging steps in the GitHub Actions workflow.
- Add separate badges for Test and Packaging in the README.

Thank you!

For review https://github.com/openjournals/joss-reviews/issues/6533
