### 2024-06-19 Sm00thix/IKPLS#27 @basileMarchand
Clarify installation instructions for GPU usage
Hello again !

In the installation section, it is mentioned that running `pip install ikpls`  is sufficient, which is true for **CPU** usage.
However, for **GPU** usage, there is *additional work* required, particularly for the installation of Jax. A user might mistakenly believe that running `pip install ikpls` on a computer with a GPU would automatically configure everything for GPU usage, which is not the case.

Could you please add a warning in the documentation to inform users that if they want to use the GPU capabilities, they need to install Jax separately and give ? Ideally, provide some guidance on how to perform this installation.

Thank you !

For review https://github.com/openjournals/joss-reviews/issues/6533
