### 2024-06-20 34a4f58
Changing workflows and adding badges in relation to #26
