+++ .github/workflows/workflow.yml
@@ -1,45 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
-name: Python package
-
-on:
-  push:
-    branches: [ "main" ]
-    paths:
-      - 'ikpls/**'
-      - 'tests/**'
-      - 'pyproject.toml'
-      - '.github/**'
-
-jobs:
-  test_package:
-    runs-on: ${{ matrix.os }}  
-    strategy:  
-      fail-fast: false  
-      matrix:  
-        os: [ubuntu-latest, windows-latest, macos-12]
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/test
-        with:
-          PYTHON_VERSION: ${{ matrix.python-version }}
-
-  build_package:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/build
-
-  publish_package:
-    needs: [build_package, test_package]
-    runs-on: ubuntu-latest
-    steps:  
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/publish
-    permissions:
-      id-token: write
-    environment:
-      name: pypi
-      url: https://pypi.org/p/ikpls
