### 2024-06-20 Sm00thix/IKPLS#27 @Sm00thix
Hi @basileMarchand!

I believe this is what I address in the **Pre-requisites** section in the README. There, I have provided a link to the [JAX Installation Guide](https://jax.readthedocs.io/en/latest/installation.html) which addresses configuration for usage with CPU, GPU, and TPU. As the JAX installation guide may change, I believe the best I can do is to refer users to the official documentation to ensure that nothing I write will be out of date.

Please let me know if I have understood correctly :-)
