### 2024-06-20 Sm00thix/IKPLS#29 @Sm00thix
Changes to paper.md introduction and first section in README
Quoting @basileMarchand from the [review at JOSS](https://github.com/openjournals/joss-reviews/issues/6533):

In the JOSS paper, it would be good in my opinion to have a more "simplified" introduction in the summary so that a less experienced reader can easily understand what need the package is addressing. Also, the first paragraph of the README (which is also the first paragraph of the documentation) should be made more engaging to catch the attention of readers who might be searching without really knowing what they need. This would help attract more users and build a more diverse community.
