### 2024-06-20 #6533 @Sm00thix
> Hello @Sm00thix
>
> first of all, sorry for the excessive delay for my review
>
> In the JOSS paper, it would be good in my opinion to have a more "simplified" introduction in the summary so that a less experienced reader can easily understand what need the package is addressing. Also, the first paragraph of the README (which is also the first paragraph of the documentation) should be made more engaging to catch the attention of readers who might be searching without really knowing what they need. This would help attract more users and build a more diverse community.

Hi @basileMarchand

No worries. Life happens :-)

Thank you for your feedback. I will make sure to incorporate it. I created an [issue](https://github.com/Sm00thix/IKPLS/issues/29) on the IKPLS repo and will address the feedback from there.
