+++ .github/workflows/pull_request_package_workflow.yml
@@ -1,15 +1,17 @@
 name: Pull Request Python Package
 
 on:
-    workflow_run:
-        workflows: ["Pull Request Tests"]
-        types:
-            - completed
+    pull_request:
+        branches: [ "main" ]
+        paths:
+            - 'ikpls/**'
+            - 'tests/**'
+            - '.github/**'
+            - 'pyproject.toml'
 
 jobs:
     build_package:
         runs-on: ubuntu-latest
-        if: ${{ github.event.workflow_run.conclusion == 'success' }}
         steps:
             - uses: actions/checkout@v4
             - uses: ./.github/actions/build
