+++ .github/workflows/package_workflow.yml
@@ -9,7 +9,6 @@ on:
 jobs:
     build_package:
         runs-on: ubuntu-latest
-        if: ${{ github.event.workflow_run.conclusion == 'success' }}
         steps:
             - uses: actions/checkout@v4
             - uses: ./.github/actions/build
