+++ .github/actions/build/action.yml
@@ -22,9 +22,9 @@ runs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 ikpls/ --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 ikpls/ --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       shell: bash
 
     - name: Install IKPLS dependencies
