### 2024-06-20 33b0327
Updated build action flake8 linting to use max line length of 88. Also did a bunch of formatting related to linting.
