+++ ikpls/numpy_ikpls.py
@@ -29,15 +29,15 @@ class PLS(BaseEstimator):
     ----------
     algorithm : int, default=1
         Whether to use Improved Kernel PLS Algorithm #1 or #2.
-    
+
     center_X : bool, default=True
         Whether to center `X` before fitting by subtracting its row of
         column-wise means from each row.
-    
+
     center_Y : bool, default=True
         Whether to center `Y` before fitting by subtracting its row of
         column-wise means from each row.
-    
+
     scale_X : bool, default=True
         Whether to scale `X` before fitting by dividing each row with the row of `X`'s
         column-wise standard deviations. Bessel's correction for the unbiased estimate
