### 2024-06-20 0c2ebab
Changing workflows and adding badges in relation to #26 (#28)
