+++ tests/test_ikpls.py
@@ -3613,7 +3613,7 @@ class TestClass:
             mse = np.mean(se, axis=-2)
             rmse = np.sqrt(mse)
             return rmse
-        
+
         def jax_rmse_per_component(
                 Y_true: npt.NDArray, Y_pred: npt.NDArray
         ) -> npt.NDArray:
@@ -3624,7 +3624,7 @@ class TestClass:
             mse = np.mean(se, axis=-2)
             rmse = jnp.sqrt(mse)
             return rmse
-        
+
         jnp_splits = jnp.array(splits)
 
         def cv_splitter(splits: npt.NDArray):
@@ -3753,14 +3753,15 @@ class TestClass:
 
             # Convert the results from dict to list for easier comparison
             fast_cv_np_pls_alg_1_results = [
-                        np.asarray(value) for value in fast_cv_np_pls_alg_1_results.values()
+                        np.asarray(value) for value
+                        in fast_cv_np_pls_alg_1_results.values()
                     ]
             fast_cv_np_pls_alg_2_results = [
                 np.asarray(value) for value in fast_cv_np_pls_alg_2_results.values()
             ]
 
-            # Sort fast cv results according to the unique splits for comparison with the
-            # other algorithms
+            # Sort fast cv results according to the unique splits for comparison with
+            # the other algorithms
             unique_splits, sort_indices = np.unique(splits, return_index=True)
             unique_splits = unique_splits.astype(int)
             fast_cv_order = np.argsort(sort_indices)
@@ -3810,15 +3811,21 @@ class TestClass:
                 ["RMSE"],
             )
 
-            # Get the best number of components in terms of minimizing validation RMSE for
-            # each split is equal among all algorithms
+            # Get the best number of components in terms of minimizing validation RMSE
+            # for each split is equal among all algorithms
             unique_splits = np.unique(splits).astype(int)
             np_pls_alg_1_best_num_components = [
-                [np.argmin(np_pls_alg_1_rmses[split][..., i]) for split in unique_splits]
+                [
+                    np.argmin(np_pls_alg_1_rmses[split][..., i])
+                    for split in unique_splits
+                ]
                 for i in range(M)
             ]
             np_pls_alg_2_best_num_components = [
-                [np.argmin(np_pls_alg_2_rmses[split][..., i]) for split in unique_splits]
+                [
+                    np.argmin(np_pls_alg_2_rmses[split][..., i])
+                    for split in unique_splits
+                ]
                 for i in range(M)
             ]
             fast_cv_np_pls_alg_1_best_num_components = [
@@ -3865,14 +3872,16 @@ class TestClass:
             ]
             np_pls_alg_1_best_rmses = [
                 [
-                    np_pls_alg_1_rmses[split][np_pls_alg_1_best_num_components[i][split], i]
+                    np_pls_alg_1_rmses[split]
+                    [np_pls_alg_1_best_num_components[i][split], i]
                     for split in unique_splits
                 ]
                 for i in range(M)
             ]
             np_pls_alg_2_best_rmses = [
                 [
-                    np_pls_alg_2_rmses[split][np_pls_alg_2_best_num_components[i][split], i]
+                    np_pls_alg_2_rmses[split]
+                    [np_pls_alg_2_best_num_components[i][split], i]
                     for split in unique_splits
                 ]
                 for i in range(M)
@@ -3932,22 +3941,49 @@ class TestClass:
                 for i in range(M)
             ]
 
-            assert_allclose(np_pls_alg_2_best_rmses, np_pls_alg_1_best_rmses, atol=atol, rtol=rtol)
             assert_allclose(
-                fast_cv_np_pls_alg_1_best_rmses, np_pls_alg_1_best_rmses, atol=atol, rtol=rtol
[…]
