+++ ikpls/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.2.post2"
+__version__ = "1.2.2.post3"
