+++ .github/workflows/package_workflow.yml
@@ -1,0 +1,27 @@
+name: Python Package
+
+on:
+    workflow_run:
+        workflows: ["Tests"]
+        types:
+            - completed
+
+jobs:
+    build_package:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v4
+            - uses: ./.github/actions/build
+
+    publish_package:
+        needs: build_package
+        runs-on: ubuntu-latest
+        if: ${{ github.event.workflow_run.conclusion == 'success' }}
+        steps:  
+            - uses: actions/checkout@v4
+            - uses: ./.github/actions/publish
+        permissions:
+            id-token: write
+        environment:
+            name: pypi
+            url: https://pypi.org/p/ikpls
