+++ .github/workflows/test_workflow.yml
@@ -1,0 +1,24 @@
+name: Tests
+
+on:
+    push:
+        branches: [ "main" ]
+        paths:
+            - 'ikpls/**'
+            - 'tests/**'
+            - '.github/**'
+            - 'pyproject.toml'
+
+jobs:
+    test_package:
+        runs-on: ${{ matrix.os }}  
+        strategy:  
+            fail-fast: false  
+            matrix:  
+                os: [ubuntu-latest, windows-latest, macos-12]
+                python-version: ["3.9", "3.10", "3.11", "3.12"]
+        steps:
+            - uses: actions/checkout@v4
+            - uses: ./.github/actions/test
+              with:
+                    PYTHON_VERSION: ${{ matrix.python-version }}
