+++ .github/workflows/pull_request_test_workflow.yml
@@ -1,7 +1,4 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
-name: Python package
+name: Pull Request Tests
 
 on:
   pull_request:
@@ -9,6 +6,7 @@ on:
     paths:
       - 'ikpls/**'
       - 'tests/**'
+      - '.github/**'
       - 'pyproject.toml'
 
 jobs:
@@ -24,9 +22,3 @@ jobs:
       - uses: ./.github/actions/test
         with:
           PYTHON_VERSION: ${{ matrix.python-version }}
-
-  build_package:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/build
