+++ .github/workflows/pull_request_package_workflow.yml
@@ -1,0 +1,17 @@
+name: Pull Request Python Package
+
+on:
+    pull_request:
+        branches: [ "main" ]
+        paths:
+            - 'ikpls/**'
+            - 'tests/**'
+            - '.github/**'
+            - 'pyproject.toml'
+
+jobs:
+    build_package:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v4
+            - uses: ./.github/actions/build
