### 2024-06-20 Sm00thix/IKPLS#26 @Sm00thix
Hi @basileMarchand,

I believe, I have resolved this issue in https://github.com/Sm00thix/IKPLS/commit/0c2ebabf7b37f0ce610899ea2749c6b612a8f0f9. Do you agree?
