### 2024-06-22 7e4d595
Updated paper.md and README.md in relation to #29
