+++ paper/paper.md
@@ -32,28 +32,26 @@ affiliations:
     index: 4
   - name: Natural History Museum of Denmark (NHMD), University of Copenhagen, Denmark
     index: 5
-date: 09 January 2024
+date: 22 June 2024
 bibliography: paper.bib
 ---
 
 # Summary
-The `ikpls` software package introduces fast and versatile implementations of both versions of Improved Kernel Partial Least Squares (IKPLS) [@dayal1997improved]. `ikpls` also contains the fast cross-validation algorithm presented by Engstrøm [@engstrøm2024shortcutting] combined with both IKPLS algorithms.
+The `ikpls` software package provides fast and efficient tools for PLS (partial least squares) modeling. This package is designed to help researchers and practitioners handle PLS modeling faster than previously possible - particularly on large datasets. The PLS implementations in `ikpls` use the fast IKPLS (Improved Kernel PLS) algorithms [@dayal1997improved], providing a substantial speedup compared to scikit-learn's [@scikit-learn] PLS implementation, which is based on NIPALS (Nonlinear Iterative Partial Least Squares) [@wold1966estimation]. The `ikpls` package also offers an implementation of IKPLS combined with the fast cross-validation algorithm by Engstrøm [@engstrøm2024shortcutting], significantly accelerating cross-validation of PLS models - especially when using a large number of cross-validation splits.
 
 `ikpls` offers NumPy-based CPU and JAX-based CPU/GPU/TPU implementations. The JAX implementations are also differentiable, allowing seamless integration with deep learning techniques. This versatility enables users to handle diverse data dimensions efficiently.
 
-Benchmarks demonstrate the superior performance of `ikpls` compared to scikit-learn's NIPALS across various scenarios. Additionally, the software addresses the redundant structure in cross-validation, proposing an algorithmic improvement for substantial speedup without recomputing total matrix products.
-
-In conclusion, `ikpls` empowers researchers and practitioners in chemometrics and related fields with efficient, scalable, and end-to-end differentiable tools for PLS modeling, facilitating optimal component selection and preprocessing decisions by offering implementations of
+In conclusion, `ikpls` empowers researchers and practitioners in machine learning, chemometrics, and related fields with efficient, scalable, and end-to-end differentiable tools for PLS modeling, facilitating optimal component selection and preprocessing decisions by offering implementations of
 
 1. both variants of IKPLS for CPUs;
 2. both variants of IKPLS for GPUs, both of which are end-to-end differentiable, allowing integration with deep learning models;
-3. a new algorithm for cross-validation that yields a substantial speedup if the training set is larger than the validation set.
+3. IKPLS combined with a cross-validation algorithm that yields a substantial speedup compared to the regular cross-validation algorithm.
 
 # Statement of need
 
[…]
