### 2024-06-22 09c2ee3
Another, minor update of paper.md and README.md in relation to #29.
