+++ README.md
@@ -36,20 +36,20 @@ The documentation is available at
 
 In addition to the standalone IKPLS implementations, this package
 contains an implementation of IKPLS combined with the novel, fast cross-validation
-algorithms by Engstrøm [[7]](#references). The fast cross-validation algorithms
+ by Engstrøm [[7]](#references). The fast cross-validation algorithm
 benefit both IKPLS Algorithms and especially Algorithm #2. The fast
-cross-validation algorithms are mathematically equivalent to the
-classical cross-validation algorithm. Still, they are much quicker.
-The fast cross-validation algorithms **correctly handle (column-wise)
+cross-validation algorithm is mathematically equivalent to the
+classical cross-validation algorithm. Still, it is much quicker.
+The fast cross-validation algorithm **correctly handles (column-wise)
 centering and scaling** of the X and Y input matrices using training set means and
 standard deviations to avoid data leakage from the validation set. This centering
 and scaling can be enabled or disabled independently from eachother and for X and Y 
 by setting the parameters `center_X`, `center_Y`, `scale_X`, and `scale_Y`, respectively.
 In addition to correctly handling (column-wise) centering and scaling,
-the fast cross-validation algorithms **correctly handle row-wise preprocessing**
+the fast cross-validation algorithm **correctly handles row-wise preprocessing**
 such as (row-wise) centering and scaling of the X and Y input matrices,
 convolution, or other preprocessing. Row-wise preprocessing can safely be
-applied before passing the data to the fast cross-validation algorithms.
+applied before passing the data to the fast cross-validation algorithm.
 
 ## Prerequisites
 
