### 2024-06-22 3fb1052
Another, minor update of paper.md in relation to #29.
