### 2024-06-22 Sm00thix/IKPLS#29 @Sm00thix
Hi @basileMarchand,

I have done my best to implement the requested changes.
Please, let me know what you think.
