### 2024-06-22 fe01ee6
Fixed typo in README.
