### 2024-06-24 Sm00thix/IKPLS#27 @boisgera
Hi @Sm00thix!

I agree with you that you give the relevant information in [the first section](https://github.com/Sm00thix/IKPLS?tab=readme-ov-file#unlock-the-power-of-fast-and-stable-partial-least-squares-modeling-with-ikpls).

But I guess that some potential users won't read everything in the README linearly but will instead skip to the first piece of code that looks like an install command (especially if they are not familiar with the JAX ecosystem and just care about having a fast implementation of the algorithm).

Maybe if you add in the prerequisite section something like :

> To enable GPU support for IKPLS, you need to install JAX with CUDA before, typically with:
> ```
> pip3 install -U "jax[cuda12]"
> ```
> Check the [JAX Installation Guide](https://jax.readthedocs.io/en/latest/installation.html) for details. [... then more about TPU, etc. ...]

that would highlight the need for the extra step? (I agree that you should not try to replicate the full JAX installation guide in your README but link to it instead!)

Or do you think that this addition would be counterproductive?
