### 2024-06-25 e318e8e
Gave more detailed instructions for GPU and TPU installation. Related to #27
