### 2024-06-25 Sm00thix/IKPLS#27 @Sm00thix
Hi @boisgera and @basileMarchand!

I have taken the suggestion by @boisgera and, additionally, also added clarification for a JAX TPU installation.
I agree that this is the most user-friendly approach. I have clarified that if a user desires a more customized JAX installation, they should follow the JAX Installation Guide. These changes have been implemented in https://github.com/Sm00thix/IKPLS/commit/e318e8e5af3f61db5f0295a34843199401ee926e.

If you agree with this, then we can consider this issue closed :-)
