### 2024-06-26 841dc4a
Added support for estimation of benchmarking for sk, np1, and np2. Also modified paper/README.md to clarify the estimation process. Also updated the notebook to reflect the newly added --estimate flag in time_pls.py. This is related to #24
