### 2024-06-26 85d8046
refactor the notebook for reproducing results
more options available from the command line
more consistent way to implement the various filtering stages
