### 2024-06-27 1474df9
reproducing notebook: replace -s (bool) with -s <n>
this way we can go for more and more complex scenarios
also the shortcut for --dry-run becomes more traditional -n
