### 2024-06-27 d2cdde5
reproducibility improvement (#30)
* clarify the estimation methods

* refactor the notebook for reproducing results

more options available from the command line
more consistent way to implement the various filtering stages

* reproducing notebook: replace -s (bool) with -s <n>

this way we can go for more and more complex scenarios
also the shortcut for --dry-run becomes more traditional -n
