+++ paper/reproducing-results-notebook.py
@@ -32,6 +32,16 @@ REF_TIMINGS = "timings/timings.csv"
 # default - see next cell for how to change it
 OUR_TIMINGS = "timings/user_timings.csv"
 
+# catch low-hanging fruits first
+SKIP_RUNS_LONGER_THAN = 0
+
+# by default, only on CPU
+RUN_ON_GPUS = False
+
+# 
+DRY_RUN = False
+
+
 # %%
 # provide a way to choose the output from the command line
 # but argparse won't work from Jupyter, so:
@@ -44,16 +54,29 @@ except:
     # not in IPython/notebook - so we run from the command line
     from argparse import ArgumentParser
     parser = ArgumentParser()
-    parser.add_argument("-o", "--output", default=OUR_TIMINGS, help="filename for the csv output")
+    parser.add_argument("-o", "--output", default=OUR_TIMINGS, 
+                        help="filename for the csv output")
+    parser.add_argument("-s", "--skip-runs-longer-than", default=SKIP_RUNS_LONGER_THAN, 
+                        action="store", type=int,
+                        help="speed up: skip runs that had taken longer than, in seconds")
+    parser.add_argument("-g", "--gpu", default=RUN_ON_GPUS,
+                        action="store_true",
+                        help="enable runs that require a GPU")
+    parser.add_argument("-n", "--dry-run", default=DRY_RUN,
+                        action="store_true",
+                        help="just show the commands to run, do not actually trigger them")
     args = parser.parse_args()
     OUR_TIMINGS = args.output
+    SKIP_RUNS_LONGER_THAN = args.skip_runs_longer_than
+    RUN_RUN_ON_GPUS = args.gpu
+    DRY_RUN = args.dry_run
 
-print(f"using {OUR_TIMINGS=}")
+print(f"using {OUR_TIMINGS=} {SKIP_RUNS_LONGER_THAN=} {RUN_ON_GPUS=} {DRY_RUN=}")
 
 # %% [markdown]
 # ## loading the paper timings
 #
-# to get a list of what would need to be done
+# to get a list of what would need to be done; the reference file here comes with the paper
 
 # %%
 import pandas as pd
@@ -63,12 +86,18 @@ paper = pd.read_csv(REF_TIMINGS)
 paper.shape
 
 # %%
-paper.head()
+paper.head(3)
+
+# %%
+total = len(paper)
 
 # %% [markdown]
 # ### dropping the njobs column
 #
 # we won't need this
+
+# %%
+paper[paper.inferred.notna() & paper.njobs.notna()]
 
 # %%
 paper.drop(columns=['njobs'], inplace=True)
@@ -77,14 +106,14 @@ paper.shape
 # %% [markdown]
 # ## loading previous runs
 #
-# to avoid re-running them if already done
+# to avoid re-running them if already done; here we load our local file, the one that contains **OUR** previous runs
 
 # %%
 try:
     previous = (
         pd.read_csv(OUR_TIMINGS)
         .rename(columns={'time': 'previous'})
-        # this is unchanged between focus and previous
+        # this is unchanged between paper and previous
         # so avoid duplication in the merge below
         .drop(columns=['inferred'])
     )
@@ -97,27 +126,7 @@ else:
     print(f"restarting from scratch")
 
 # %% [markdown]
-# ## running again
-#
-# focus is a selection of the runs to reproduce - e.g. we will use it optionnally later on if we can only run on a CPU
-
-# %%
-focus = paper.copy()
-
-def status():
-    print(f"we are focusing on {len(focus)} runs")
-
[…]
