### 2024-06-27 4f1e777
Clarified instructions for manual benchmarking in paper/README.md. Related to #24
