### 2024-06-27 ee3de3e
Benchmarking clarifications (#31)
* Added support for estimation of benchmarking for sk, np1, and np2. Also modified paper/README.md to clarify the estimation process. Also updated the notebook to reflect the newly added --estimate flag in time_pls.py. This is related to #24

* reproducibility improvement (#30)

* clarify the estimation methods

* refactor the notebook for reproducing results

more options available from the command line
more consistent way to implement the various filtering stages

* reproducing notebook: replace -s (bool) with -s <n>

this way we can go for more and more com
[…]
