### 2024-06-27 822ce3a
reproducing notebook: fix the -g option
before that change, was always skipping gpu runs
also the default is now to run all, use -g to skip the runs that require a GPU
