### 2024-06-27 cb2ab1f
reproduce: configure jax to enable 64 bits
