### 2024-06-27 9cf6cc7
Fixed FutureDeprecationWarning arising from hashing of a tracer. This was hashing of the value i in PLSBase._step_3
