### 2024-06-27 2a47921
Updated workflows to run tests on latest macos instead of macos-12.
