+++ ikpls/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.2.post3"
+__version__ = "1.2.3"
