### 2024-06-28 97c9024
Updated workflows to fall back to macos-12. Also fixed error for reverse mode differentiable JAX.
