+++ .github/workflows/test_workflow.yml
@@ -15,7 +15,7 @@ jobs:
         strategy:  
             fail-fast: false  
             matrix:  
-                os: [ubuntu-latest, windows-latest, macos-latest]
+                os: [ubuntu-latest, windows-latest, macos-12]
                 python-version: ["3.9", "3.10", "3.11", "3.12"]
         steps:
             - uses: actions/checkout@v4
