### 2024-06-28 d5a5fb6
Dev (#32)
* Fixed FutureDeprecationWarning arising from hashing of a tracer. This was hashing of the value i in PLSBase._step_3

* Updated workflows to run tests on latest macos instead of macos-12.

* Bumped ikpls version number

* Updated workflows to fall back to macos-12. Also fixed error for reverse mode differentiable JAX.
