### 2024-06-28 Sm00thix/IKPLS#26 @basileMarchand
Hi @Sm00thix,

That's perfect! Personally, I find accessing the results of Actions much easier now.
