### 2024-06-28 Sm00thix/IKPLS#29 @basileMarchand
Hi  @Sm00thix,

thank you for the work on rewriting the introduction. I think it makes the paper more engaging now because it's more accessible to people who are not specialists (they can understand what they just came across). I believe that even for you, for the dissemination of your package, this will help reach more people. Just one last suggestion, I think you could take the first 2-3 sentences of the JOSS paper summary and integrate them directly into the README, in the first section, so that anyone who comes across the IKPLS documentation or repository can immediately understand what they have found.
