### 2024-06-29 36b10ae
Extended README.md with the first to sentences of paper/paper.md. Also fixed some capitalization in paper/paper.md. This is related to #29
