### 2024-06-29 Sm00thix/IKPLS#29 @Sm00thix
Hi @basileMarchand!

Thank you for your suggestions and aid in improving the reach of `ikpls`. I have incorporated your last suggestion in https://github.com/Sm00thix/IKPLS/commit/36b10aedb42de483d89c3b9adf24723840401439. :-)
