### 2024-06-29 b27201b
a few convenience tweaks to make the reproduce script smoother (#34)
* reproducing notebook: fix the -g option

before that change, was always skipping gpu runs
also the default is now to run all, use -g to skip the runs that require a GPU

* reproduce: configure jax to enable 64 bits

* cosmetic
