### 2024-07-07 #6533 @boisgera
Hi everyone!

I see that @parmentelat has checked all the items from his list ; there are still a few unchecked items for @basileMarchand. Anything I can do to help about these items?

All issues opened so far have been adressed and closed, except for https://github.com/Sm00thix/IKPLS/issues/29.
@basileMarchand, did the change https://github.com/Sm00thix/IKPLS/commit/36b10aedb42de483d89c3b9adf24723840401439 address your concern? Can you close the issue?

Kind regards,

Sébastien
