### 2024-07-07 f1a60ea
Now comparing std to 0 using float epsilon instead of direct comparison. This seems to have been the cause of #33
