### 2024-07-07 9ccc4e3
Updated tolerance in test to account for machine epsilon in float comparison
