### 2024-07-07 0ddea86
Compare std deviation from 0 with float epsilon instead of exact equality between std and 0 (#35)
