### 2024-07-12 #6533 @basileMarchand
Hi @boisgera, @parmentelat @Sm00thix

I have now completed my review and now checked all the boxes.

I have two general comments to make before closing my contribution to this review:

**Performance Evaluation:**

Unfortunately, I was unable to evaluate the performance on GPU as I could not successfully install JAX. Thus, my review only covers the CPU version.

**References:**

The references cited in the paper appear to be relevant. However, I am not an expert in this specific field, so my judgment is limited to the extent of my current knowledge.

Thank you for the opportunity to review this submission.

Basile
