### 2024-07-22 #6533 @boisgera
Hi everyone,

Everything looks good at this stage, thanks @parmentelat, @basileMarchand and @Sm00thix for your efforts 🙏
We are now entering the "post-review" stage!

@Sm00thix can you make a new release of `ikpls` (if it's necessary/appropriate)  and deposit a new archive of the software with Zenodo/figshare? I'll need the corresponding archive DOI to proceed.

Cheers

Sébastien
