### 2024-07-22 #6533 @boisgera
Hi @Sm00thix,

It seems that there are a small number of minor issues with the current Zenodo archive:

 - [ ] The title of the archive does not match the (full) title of the JOSS paper,
 - [ ] The list of authors does not match the declared list of authors for the JOSS paper (and the ORCID tags are not included),
 - [ ] The license of the archive does not match the sofware license.

See e.g. https://joss.theoj.org/papers/10.21105/joss.06566 and the archive https://zenodo.org/records/12775561 to see how these metadata should match.

I guess that these metadata are immutable and that you will need to make a new Zenodo archive (?).
Sorry for the inconvenience...
