### 2024-07-22 #6533 @Sm00thix
Hi @boisgera,

I apologize for the oversights. I have taken care of all three points. Interestingly, I was able to edit the metadata without changing the DOI. In fact, Zenodo notified me that I can not change the files in the archive, but I can always change the metadata.

Please let me know if I need to do anything else! :-)
