### 2024-07-23 #6533 @crvernon
:wave: @Sm00thix - I just need you to make the following changes before we can accept this for publication:

**In your paper:**

- [ ] LINE 18: "...algorithm by Engstrøm (O.-C. G. Engstrøm, 2024), significantly..." should use an in text citation resulting in "algorithm by Engstrøm (2024), significantly...".  You can achieve this by removing the brackets from the in-text mention of the bib reference in your paper.  For example, instead of `by Engstrøm [@ref]` you can simply use `by @ref`
- [ ] LINE 34-35:  The references for S. Wold should appear together in a single set of parenthesis.  This formatting issue occurs in several other places throughout the paper.  I will not mention all of them, but please fix those as well.  To fix this, just combine the references in a single set of brackets delimited by a semicolon (e.g., `[@ref1; @ref2]`
- [ ] LINE 34-36:  I am also confused by the phrasing here "...or a classification model, PLS-DA (PLS discriminant analysis), (Barker & Rayens, 2003)..." should this be "...or a classification model, PLS-DA (PLS discriminant analysis) (Barker & Rayens, 2003)..." where the comma is removed before the reference?  If so, please remove other similar occurrences of this in the text.
- [ ] LINE 47:  Same comment here about in-line text citation for the Dayal and MacGregor citation.
- [ ] There are many missing DOI in your references.  Please add them into your bib file.  See the message generated before this for hints but you will need to confirm that these are correct.
- [ ] Ensure that items needing capitalization in the references are accounted for.  You can maintain formatting in your references using curly brackets around parts of text in your bib file that you want keep the formatting of.
