### 2024-07-23 b05ae1d
Updated paper in response to https://github.com/openjournals/joss-reviews/issues/6533
