+++ paper/paper.bib
@@ -22,7 +22,8 @@
   number={2},
   pages={109--130},
   year={2001},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s0169-7439(01)00155-1}
 }
 
 @article{barker2003partial,
@@ -33,7 +34,8 @@
   number={3},
   pages={166--173},
   year={2003},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/cem.785}
 }
 
 @article{dayal1997improved,
@@ -44,7 +46,8 @@
   number={1},
   pages={73--85},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/(sici)1099-128x(199701)11:1<73::aid-cem435>3.0.co;2-#}
 }
 
 @article{scikit-learn,
@@ -67,7 +70,8 @@
   number={10},
   pages={1201--1222},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.trac.2009.07.007}
 }
 
 @article{sorensen2021nir,
@@ -76,7 +80,8 @@
   journal={Near-Infrared Spectroscopy: Theory, Spectral Analysis, Instrumentation, and Applications},
   pages={127--189},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-981-15-8648-4_7}
 }
 
 @article{du2022quantitative,
@@ -87,7 +92,8 @@
   number={3},
   pages={2956--3009},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/1541-4337.12867}
 }
 
 @article{alin2009comparison,
@@ -98,7 +104,8 @@
   number={4},
   pages={711},
   year={2009},
-  publisher={Springer Nature BV}
+  publisher={Springer Nature BV},
+  doi={10.1007/s00362-008-0123-6}
 }
 
 @article{andersson2009comparison,
@@ -109,7 +116,8 @@
   number={10},
   pages={518--529},
   year={2009},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/cem.1248}
 }
 
 @inproceedings{engstrom2023improving,
@@ -117,14 +125,16 @@
   author={Engstr{\o}m, O.-C. G. and Dreier, E. S. and Jespersen, B. M. and Pedersen, K. S.},
   booktitle={Proceedings of the IEEE/CVF International Conference on Computer Vision},
   pages={485--494},
-  year={2023}
+  year={2023},
+  doi={10.1109/iccvw60793.2023.00055}
 }
 
 @article{engstrom2023analyzing,
   title={Analyzing Near-Infrared Hyperspectral Imaging for Protein Content Regression and Grain Variety Classification Using Bulk References and Varying Grain-to-Background Ratios},
   author={Engstr{\o}m, O.-C. G. and Dreier, E. S. and Jespersen, B. M. and Pedersen, K. S.},
   journal={arXiv preprint arXiv:2311.04042},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2311.04042}
 }
 
 @article{harris2020array,
@@ -135,7 +145,8 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2}
 }
 
 @software{jax2018github,
@@ -163,7 +174,8 @@
   number={3},
   pages={107--121},
   year={2022},
[…]
