### 2024-07-23 45a900c
Fixed wrong DOIs in response to https://github.com/openjournals/joss-reviews/issues/6533
