+++ paper/paper.bib
@@ -47,7 +47,7 @@
   pages={73--85},
   year={1997},
   publisher={Wiley Online Library},
-  doi={10.1002/(sici)1099-128x(199701)11:1<73::aid-cem435>3.0.co;2-#}
+  doi={10.1002/(sici)1099-128x(199701)11:1<73::aid-cem435>3.0.co;2-\#}
 }
 
 @article{scikit-learn,
@@ -93,7 +93,7 @@
   pages={2956--3009},
   year={2022},
   publisher={Wiley Online Library},
-  doi={10.1111/1541-4337.12867}
+  doi={10.1111/1541-4337.12958}
 }
 
 @article{alin2009comparison,
@@ -105,7 +105,7 @@
   pages={711},
   year={2009},
   publisher={Springer Nature BV},
-  doi={10.1007/s00362-008-0123-6}
+  doi={10.1007/s00362-009-0251-7}
 }
 
 @article{andersson2009comparison,
