+++ paper/paper.bib
@@ -47,7 +47,7 @@
   pages={73--85},
   year={1997},
   publisher={Wiley Online Library},
-  doi={10.1002/(sici)1099-128x(199701)11:1<73::aid-cem435>3.0.co;2-\#}
+  doi={10.1002/(SICI)1099-128X(199701)11:1<73::AID-CEM435>3.0.CO;2-\#}
 }
 
 @article{scikit-learn,
