### 2024-07-23 f1585cc
Another attempt to fix DOI for dayal1997improved in response to https://github.com/openjournals/joss-reviews/issues/6533
