### 2024-07-23 103a87c
Yet another attempt to fix DOI for dayal1997improved in response to https://github.com/openjournals/joss-reviews/issues/6533
