### 2024-07-23 #6533 @Sm00thix
Hi @crvernon!

Thanks for the feedback. I have implemented the changes as per your instructions. Do I need to create a new archive on Zenodoto reflect the changes in paper/paper.md and paper/paper.bib?

> 👋 @Sm00thix - I just need you to make the following changes before we can accept this for publication:
>
> **In your paper:**
>
> * [ ]  LINE 18: "...algorithm by Engstrøm (O.-C. G. Engstrøm, 2024), significantly..." should use an in text citation resulting in "algorithm by Engstrøm (2024), significantly...".  You can achieve this by removing the brackets from the in-text mention of the bib reference in your paper.  For example, instead of `by Engstrøm [@ref]` you can simply use `by @ref`
Done.
> * [ ]  LINE 34-35:  The references for S. Wold should appear together in a single set of parenthesis.  This formatting issue occurs in several other places throughout the paper.  I will not mention all of them, but please fix those as well.  To fix this, just combine the references in a single set of brackets delimited by a semicolon (e.g., `[@ref1; @ref2]`
Done.
> * [ ]  LINE 34-36:  I am also confused by the phrasing here "...or a classification model, PLS-DA (PLS discriminant analysis), (Barker & Rayens, 2003)..." should this be "...or a classification model, PLS-DA (PLS discriminant analysis) (Barker & Rayens, 2003)..." where the comma is removed before the reference?  If so, please remove other similar occurrences of this in the text.
Yes, the comma was a mistake. I removed it. I found no other similar occurences in the text.
> * [ ]  LINE 47:  Same comment here about in-line text citation for the Dayal and MacGregor citation.
Done.
> * [ ]  There are many missing DOI in your references.  Please add them into your bib file.  See the message generated before this for hints but you will need to confirm that these are correct.
I have added the DOIs for the references that have DOIs.
> * [ ]  Ensure that items needing capitalization in the references are accounted for.  You can maintain formatting in your references using curly brackets around parts of text in your bib file that you want keep the formatting of.
I am not entirely sure what you mean by this. Do I need to change anything? I am satisfied with the capitalization in the references, if this is what you are asking.

I also deleted lines 75-78 as these were simply duplicating what was stated in 70-74. I think I had accidentally mixed two versions of the same piece of text.
