### 2024-07-23 #6533 @crvernon
A note for the thread:

> INVALID DOIs
> - 10.1002/(SICI)1099-128X(199701)11:1<73::AID-CEM435>3.0.CO;2-23 is INVALID

This is actually a valid DOI, it just renders in the check without the `%` but is correctly represented in the paper.
