### 2024-07-23 #6533 @Sm00thix
Hi @crvernon,

Thanks for publishing the paper!

I have noticed that the paper does not show on the [[frontpage](https://joss.theoj.org/papers/10.21105/joss.06533)](https://joss.theoj.org/papers/10.21105/joss.06533).
It does, however, show correctly on [https://www.theoj.org/joss-papers/joss.06533/10.21105.joss.06533.pdf](https://www.theoj.org/joss-papers/joss.06533/10.21105.joss.06533.pdf).

Is this to be expected for some amount of time or is there an actual issue? And if so, is there anything I can do to fix it?

Best,
Ole
