### 2024-07-24 #6533 @boisgera
Hi @Sm00thix,

I can see the paper on both links (maybe that wasn't true 9 hours ago).

Congratulations! 😊

Cheers,
Sébastien
